Split mod_p4.rs into 4 clean function-boundary parts to clear max-lines (2954 -> 726+757+721+750)
