+++ crates/almide-mir/src/lower/mod.rs
@@ -1993,6 +1993,9 @@ include!("inline_scalar_fns.rs");
 include!("mod_p2.rs");
 include!("mod_p3.rs");
 include!("mod_p4.rs");
+include!("mod_p4_b.rs");
+include!("mod_p4_c.rs");
+include!("mod_p4_d.rs");
 include!("mod_p5.rs");
 // The desugar family (formerly one 4.8k-line mod_p6.rs), split by concern:
 include!("desugar.rs");
