+++ crates/almide-mir/src/lower/mod_p4.rs
@@ -724,2230 +724,3 @@ fn interp_list_to_string(inner: &Ty) -> (&'static str, &'static str) {
         _ => ("list", "to_string_x"),
     }
     }
-
-/// `${Option[T]}` interp routing per payload type. Verbatim text move (#781).
-fn interp_option_to_string(inner: &Ty) -> (&'static str, &'static str) {
-    use almide_lang::types::constructor::TypeConstructorId;
-    match inner {
-        Ty::Int => ("option", "to_string"),
-        Ty::String => ("option", "to_string_s"),
-        Ty::Float => ("option", "to_string_f"),
-        Ty::Bool => ("option", "to_string_b"),
-        // `${Option[List[Int]]}` → `some([1, 2, 3])` / `none` — the inner list renders like
-        // `${list}`, wrapped in `some(…)`. A deeper element routes to the UNLINKED `_x`.
-        Ty::Applied(TypeConstructorId::List, e) if e.len() == 1 && matches!(e[0], Ty::Int) => {
-            ("option", "to_string_li")
-        }
-        Ty::Applied(TypeConstructorId::List, e) if e.len() == 1 && matches!(e[0], Ty::String) => {
-            ("option", "to_string_ls")
-        }
-        Ty::Applied(TypeConstructorId::Option, e) if e.len() == 1 && matches!(e[0], Ty::Int) => {
-            ("option", "to_string_oi")
-        }
-        Ty::Applied(TypeConstructorId::Option, e) if e.len() == 1 && matches!(e[0], Ty::Bool) => {
-            ("option", "to_string_ob")
-        }
-        Ty::Applied(TypeConstructorId::Option, e) if e.len() == 1 && matches!(e[0], Ty::String) => {
-            ("option", "to_string_os")
-        }
-        Ty::Applied(TypeConstructorId::List, e) if e.len() == 1 && matches!(e[0], Ty::Bool) => {
-            ("option", "to_string_lb")
-        }
-        Ty::Applied(TypeConstructorId::List, e) if e.len() == 1 && matches!(e[0], Ty::Float) => {
-            ("option", "to_string_lf")
-        }
-        Ty::Applied(TypeConstructorId::Map, e)
-            if e.len() == 2 && matches!(e[0], Ty::String) && matches!(e[1], Ty::Int) =>
-        {
-            ("option", "to_string_msi")
-        }
-        Ty::Applied(TypeConstructorId::Result, e)
-            if e.len() == 2 && matches!(e[0], Ty::Int) && matches!(e[1], Ty::String) =>
-        {
-            ("option", "to_string_ri")
-        }
-        Ty::Applied(TypeConstructorId::Result, e)
-            if e.len() == 2 && matches!(e[0], Ty::String) && matches!(e[1], Ty::String) =>
-        {
-            ("option", "to_string_rs")
-        }
-        Ty::Applied(TypeConstructorId::Option, e)
-            if e.len() == 1
-                && matches!(&e[0], Ty::Applied(TypeConstructorId::Option, e2)
-                    if e2.len() == 1 && matches!(e2[0], Ty::Int)) =>
-        {
-            ("option", "to_string_ooi")
-        }
-        Ty::Applied(TypeConstructorId::Option, e)
-            if e.len() == 1
-                && matches!(&e[0], Ty::Applied(TypeConstructorId::List, e2)
-                    if e2.len() == 1 && matches!(e2[0], Ty::Int)) =>
-        {
-            ("option", "to_string_ooli")
-        }
-        Ty::Applied(TypeConstructorId::Result, e)
-            if e.len() == 2
-                && matches!(&e[0], Ty::Applied(TypeConstructorId::List, e2)
-                    if e2.len() == 1 && matches!(e2[0], Ty::Int))
-                && matches!(e[1], Ty::String) =>
-        {
-            ("option", "to_string_rli")
-        }
-        _ => ("option", "to_string_x"),
-    }
-    }
-
-/// `${Result[T, E]}` interp routing per (ok, err) pair. Verbatim text move (#781).
-fn interp_result_to_string(ok: &Ty, err: &Ty) -> (&'static str, &'static str) {
-    use almide_lang::types::constructor::TypeConstructorId;
-        match (ok, err) {
-            (Ty::Int, Ty::String) => ("result", "to_string"),
-            (Ty::String, Ty::String) => ("result", "to_string_ss"),
-            // `${Result[List[Int], String]}` → `ok([1, 2, 3])` / `err("<quoted>")`.
-            (Ty::Applied(TypeConstructorId::List, e), Ty::String)
-                if e.len() == 1 && matches!(e[0], Ty::Int) =>
-            {
-                ("result", "to_string_li")
-            }
-            (Ty::Applied(TypeConstructorId::List, e), Ty::String)
-                if e.len() == 1 && matches!(e[0], Ty::String) =>
-            {
-                ("result", "to_string_ls")
-            }
-            (Ty::Bool, Ty::String) => ("result", "to_string_b"),
-            (Ty::Applied(TypeConstructorId::Option, e), Ty::String)
-                if e.len() == 1 && matches!(e[0], Ty::Int) =>
-            {
-                ("result", "to_string_oi")
-            }
-            (Ty::Applied(TypeConstructorId::Option, e), Ty::String)
-                if e.len() == 1 && matches!(e[0], Ty::String) =>
-            {
-                ("result", "to_string_os")
-            }
-            (Ty::Applied(TypeConstructorId::Result, e), Ty::String)
-                if e.len() == 2 && matches!(e[0], Ty::Int) && matches!(e[1], Ty::String) =>
-            {
-                ("result", "to_string_ri")
-            }
-            (Ty::Applied(TypeConstructorId::List, e), Ty::String)
-                if e.len() == 1 && matches!(e[0], Ty::Bool) =>
-            {
-                ("result", "to_string_lb")
-            }
-            (Ty::Float, Ty::String) => ("result", "to_string_f"),
-            (Ty::Applied(TypeConstructorId::List, e), Ty::String)
-                if e.len() == 1 && matches!(e[0], Ty::Float) =>
-            {
-                ("result", "to_string_lf")
-            }
-            (Ty::Applied(TypeConstructorId::Option, e), Ty::String)
-                if e.len() == 1
-                    && matches!(&e[0], Ty::Applied(TypeConstructorId::List, e2)
-                        if e2.len() == 1 && matches!(e2[0], Ty::String)) =>
-            {
-                ("result", "to_string_osl")
-            }
-            (Ty::Applied(TypeConstructorId::Map, e), Ty::String)
-                if e.len() == 2 && matches!(e[0], Ty::String) && matches!(e[1], Ty::Int) =>
-            {
-                ("result", "to_string_msi")
-            }
-            (Ty::Applied(TypeConstructorId::Option, e), Ty::String)
-                if e.len() == 1
-                    && matches!(&e[0], Ty::Applied(TypeConstructorId::List, e2)
-                        if e2.len() == 1 && matches!(e2[0], Ty::Int)) =>
-            {
-                ("result", "to_string_oli")
-            }
-            _ => ("result", "to_string_x"),
-        }
-}
-
-/// Does a record/tuple/list/scalar VALUE of type `ty` materialize with REAL slots the recursive
-/// Display can read — the STATIC (IR-type-only) predicate the gate and the lowering BOTH consult so
-/// they agree on expand-vs-wrap BY CONSTRUCTION (no runtime-`materialized_aggregates` divergence).
-/// Matches exactly what the construction path materializes:
-///   - Int/Bool/Float/String          → yes (scalar / single heap leaf)
-///   - List[scalar]                    → yes (scalar-element block); List[heap] → NO (not materialized)
-///   - a registered record/tuple whose every field is itself `field_displayable` → yes (the
-///     nested-aggregate construction admits a SCALAR-ONLY nested block; a heap-IN-nested field would
-///     leak under the single-level mask, so it is NO)
-///   - Map/Set/Option/Result/variant/unresolved → NO
-fn field_displayable(ty: &Ty, registry: &RecordLayouts) -> bool {
-    use almide_lang::types::constructor::TypeConstructorId;
-    match ty {
-        Ty::Int | Ty::Bool | Ty::Float | Ty::String => true,
-        Ty::Applied(TypeConstructorId::List, a) if a.len() == 1 => !is_heap_ty(&a[0]),
-        Ty::Record { .. } | Ty::Tuple(_) | Ty::Named(..) => match resolve_aggregate(ty, registry) {
-            // A NESTED aggregate must be SCALAR-ONLY (the construction's `lower_owned_heap_field`
-            // admits only a scalar-only nested block — a heap-in-nested field would leak).
-            Some((_, _, fields)) => fields.iter().all(|(_, t)| !is_heap_ty(t)),
-            None => false,
-        },
-        _ => false,
-    }
-}
-
-/// Is a record/tuple interpolation PART statically EXPAND-foldable — i.e. the lowering will
-/// materialize it and read its real slots? True iff the part expr is a `Var` (a materialized
-/// aggregate binding; a literal/call result is not a tracked block) AND every field of the
-/// (resolvable) aggregate is `field_displayable`. The gate and the lowering both gate on THIS, so
-/// the synthetic-call count the gate credits equals the calls the lowering emits — for both the
-/// EXPAND path (recursive tree) and the WALL path (one `compound.to_string`).
-pub(crate) fn aggregate_part_expandable(expr: &IrExpr, registry: &RecordLayouts) -> bool {
-    if !matches!(expr.kind, IrExprKind::Var { .. }) {
-        return false; // a literal `${P{..}}` / a call `${f()}` is not a tracked materialized block
-    }
-    match resolve_aggregate(&expr.ty, registry) {
-        Some((_, _, fields)) => fields.iter().all(|(_, t)| field_displayable(t, registry)),
-        None => false,
-    }
-}
-
-/// Build the String-producing LEAF for ONE interpolation part, by type:
-///   - a literal text part → a `LitStr` (no call),
-///   - a String-typed part → the expr itself (identity, no call),
-///   - an EXPAND-foldable RECORD/TUPLE part (a materialized Var with displayable fields) → the
-///     recursive layout-driven Display ([`display_aggregate`]), an INLINE `ConcatStr` tree of
-///     per-field formatters; a NON-expandable record/tuple part → ONE unlinked `compound.to_string`
-///     wrapper (the function walls at render — never a wrong byte),
-///   - any other part with a pure `module.to_string` → `module.to_string(expr)`.
-/// Returns `None` for a part whose type has no admitted Display at all (an unresolved type) — the
-/// caller then declines the whole desugar.
-fn interp_part_leaf(p: &IrStringPart, registry: &RecordLayouts) -> Option<IrExpr> {
-    match p {
-        IrStringPart::Lit { value } => Some(lit_str(value)),
-        IrStringPart::Expr { expr } if matches!(expr.ty, Ty::String) => Some(expr.clone()),
-        // A record/tuple part: EXPAND if the lowering will materialize it; else wrap in the
-        // unlinked `compound.to_string` so the function walls (the SAME decision the gate makes).
-        IrStringPart::Expr { expr }
-            if matches!(expr.ty, Ty::Record { .. } | Ty::Tuple(_) | Ty::Named(..))
-                && resolve_aggregate(&expr.ty, registry).is_some() =>
-        {
-            // An ANONYMOUS record ALWAYS takes the generated sorted-field repr: v0 sorts
-            // anon fields by name, while the inline display_aggregate expansion reads the
-            // STRUCTURAL (source) order — expanding it would emit wrong bytes.
-            if let Ty::Record { fields } = &expr.ty {
-                // An ANONYMOUS record part — route to the generated
-                // `__repr_anonrec_<hash>` (sorted-field render); an unemitted shape
-                // (a nested payload) leaves the call unlinked = the honest wall.
-                Some(IrExpr {
-                    kind: IrExprKind::Call {
-                        target: CallTarget::Named {
-                            name: sym(&format!(
-                                "__repr_{}",
-                                crate::lower::anon_record_drop_name(fields)
-                            )),
-                        },
-                        args: vec![expr.clone()],
-                        type_args: Vec::new(),
-                    },
-                    ty: Ty::String,
-                    span: None,
-                    def_id: None,
-                })
-            } else if aggregate_part_expandable(expr, registry) {
-                display_aggregate(expr, &expr.ty, registry)
-            } else if let Ty::Named(name, _) = &expr.ty {
-                // A NAMED record outside the inline-expand subset (a recursive record, a
-                // List[record] field — the compound_repr class): route to the GENERATED
-                // `__repr_rec_<R>` (render-pipeline-injected). A record the generator does
-                // not emit leaves the call unlinked — the same honest render wall the
-                // `compound.to_string` wrapper gives, with the SAME call-count (one node).
-                Some(IrExpr {
-                    kind: IrExprKind::Call {
-                        target: CallTarget::Named {
-                            name: sym(&format!(
-                                "__repr_rec_{}",
-                                crate::lower::drop_fn_ident(name.as_str())
-                            )),
-                        },
-                        args: vec![expr.clone()],
-                        type_args: Vec::new(),
-                    },
-                    ty: Ty::String,
-                    span: None,
-                    def_id: None,
-                })
-            } else {
-                Some(to_string_call("compound", "to_string", expr.clone()))
-            }
-        }
-        // A custom-VARIANT part (`"${Overflow(\"x\")}"` / a bound variant var): route
-        // to the GENERATED `__repr_<V>` (render-pipeline-injected; the classify gate
-        // counts the same call node). A variant the generator does not emit (a field
-        // outside Int/Bool/String/nested-variant) leaves an unlinked call — the same
-        // honest render wall the `compound.to_string` wrapper gives records.
-        IrStringPart::Expr { expr }
-            if matches!(&expr.ty, Ty::Named(..))
-                && resolve_aggregate(&expr.ty, registry).is_none() =>
-        {
-            let Ty::Named(name, targs) = &expr.ty else { unreachable!() };
-            // A GENERIC-variant instance (`${l}` over `ReprEither[Int, String]`) takes
-            // the INSTANTIATION-KEYED repr — the exact key the generator derives
-            // (`repr_inst_ident`), so the call links iff the instantiation is emitted.
-            let rname = if targs.is_empty() {
-                format!("__repr_{}", crate::lower::drop_fn_ident(name.as_str()))
-            } else {
-                format!("__repr_{}", crate::lower::repr_inst_ident(name.as_str(), targs))
-            };
-            Some(IrExpr {
-                kind: IrExprKind::Call {
-                    target: CallTarget::Named { name: sym(&rname) },
-                    args: vec![expr.clone()],
-                    type_args: Vec::new(),
-                },
-                ty: Ty::String,
-                span: None,
-                def_id: None,
-            })
-        }
-        // A CONTAINER of a named record/variant (`${pts}` over List[Point], `${op}` over
-        // Option[Point], `${shapes}` over List[Shape]) — route to the GENERATED container
-        // repr (`__repr_list_rec_<R>` / `__repr_opt_rec_<R>` / `__repr_list_<V>`, emitted by
-        // `generate_variant_repr_sources` for exactly the container/element pairs collected
-        // from interp parts). An unemitted pair leaves the call unlinked — the honest wall
-        // (same contract as the bare `__repr_rec_<R>` arm above).
-        IrStringPart::Expr { expr }
-            if container_repr_name(&expr.ty, registry).is_some() =>
-        {
-            let name = container_repr_name(&expr.ty, registry).unwrap();
-            Some(IrExpr {
-                kind: IrExprKind::Call {
-                    target: CallTarget::Named { name: sym(&name) },
-                    args: vec![expr.clone()],
-                    type_args: Vec::new(),
-                },
-                ty: Ty::String,
-                span: None,
-                def_id: None,
-            })
-        }
-        // `${List[(Int, String)]}` / `${Option[(Bool, Bool)]}` — SCALAR-component
-        // tuple containers (list.sort / list.min/max results, the C-053 class): route
-        // to the generated `__repr_list_tup_<key>` / `__repr_opt_tup_<key>` walkers.
-        // A component outside Int/Bool/String keeps the existing `_x` wall routing.
-        // `(String, Int)` is EXCLUDED on the List side — `${List[(String, Int)]}`
-        // keeps its established `list.to_string_lsi` route (string_rle).
-        IrStringPart::Expr { expr }
-            if matches!(&expr.ty,
-                Ty::Applied(almide_lang::types::constructor::TypeConstructorId::List, a)
-                    if a.len() == 1
-                        && matches!(&a[0], Ty::Tuple(ts)
-                            if crate::lower::tuple_repr_ident(ts).is_some()
-                                && !(ts.len() == 2
-                                    && matches!(ts[0], Ty::String)
-                                    && matches!(ts[1], Ty::Int)))) =>
-        {
-            let Ty::Applied(_, a) = &expr.ty else { unreachable!() };
-            let Ty::Tuple(ts) = &a[0] else { unreachable!() };
-            let key = crate::lower::tuple_repr_ident(ts).unwrap();
-            Some(IrExpr {
-                kind: IrExprKind::Call {
-                    target: CallTarget::Named { name: sym(&format!("__repr_list_tup_{key}")) },
-                    args: vec![expr.clone()],
-                    type_args: Vec::new(),
-                },
-                ty: Ty::String,
-                span: None,
-                def_id: None,
-            })
-        }
-        IrStringPart::Expr { expr }
-            if matches!(&expr.ty,
-                Ty::Applied(almide_lang::types::constructor::TypeConstructorId::Option, a)
-                    if a.len() == 1
-                        && matches!(&a[0], Ty::Tuple(ts)
-                            if crate::lower::tuple_repr_ident(ts).is_some())) =>
-        {
-            let Ty::Applied(_, a) = &expr.ty else { unreachable!() };
-            let Ty::Tuple(ts) = &a[0] else { unreachable!() };
-            let key = crate::lower::tuple_repr_ident(ts).unwrap();
-            Some(IrExpr {
-                kind: IrExprKind::Call {
-                    target: CallTarget::Named { name: sym(&format!("__repr_opt_tup_{key}")) },
-                    args: vec![expr.clone()],
-                    type_args: Vec::new(),
-                },
-                ty: Ty::String,
-                span: None,
-                def_id: None,
-            })
-        }
-        // `${[a]}` over a List[<STRUCTURAL record>] (an inferred literal element — the
-        // r5 C-072 class): route to the generated `__repr_list_anonrec_<hash>` walker
-        // (its element repr renders NOMINALLY when the shape resolves to a declared
-        // record, sorted-anonymous otherwise). An unemitted shape (a nested payload)
-        // leaves the call unlinked — the honest wall.
-        IrStringPart::Expr { expr }
-            if matches!(&expr.ty,
-                Ty::Applied(almide_lang::types::constructor::TypeConstructorId::List, a)
-                    if a.len() == 1 && matches!(a[0], Ty::Record { .. })) =>
-        {
-            let Ty::Applied(_, a) = &expr.ty else { unreachable!() };
-            let Ty::Record { fields } = &a[0] else { unreachable!() };
-            Some(IrExpr {
-                kind: IrExprKind::Call {
-                    target: CallTarget::Named {
-                        name: sym(&format!(
-                            "__repr_list_{}",
-                            crate::lower::anon_record_drop_name(fields)
-                        )),
-                    },
-                    args: vec![expr.clone()],
-                    type_args: Vec::new(),
-                },
-                ty: Ty::String,
-                span: None,
-                def_id: None,
-            })
-        }
-        IrStringPart::Expr { expr } => {
-            let (module, func) = interp_to_string_call(&expr.ty)?;
-            Some(to_string_call(module, func, expr.clone()))
-        }
-    }
-}
-
-/// The generated container-repr callee for a `${List[<record>]}` / `${Option[<record>]}` /
-/// `${List[<variant>]}` interp part, or `None` for every other type (falls to the
-/// `interp_to_string_call` table). Record-vs-variant discrimination mirrors the bare-part
-/// arms: a `Named` that resolves in the record registry is a record, else a variant.
-fn container_repr_name(ty: &Ty, registry: &RecordLayouts) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId;
-    let named = |t: &Ty| -> Option<(String, bool)> {
-        let Ty::Named(n, _) = t else { return None };
-        Some((n.as_str().to_string(), resolve_aggregate(t, registry).is_some()))
-    };
-    match ty {
-        Ty::Applied(TypeConstructorId::List, a) if a.len() == 1 => {
-            // A generic-variant INSTANTIATION element (`${forest}` over
-            // `List[Tree[Int]]` — C-010): the instantiation-KEYED walker.
-            if let Ty::Named(n, args) = &a[0] {
-                if !args.is_empty() && resolve_aggregate(&a[0], registry).is_none() {
-                    return Some(format!(
-                        "__repr_list_{}",
-                        crate::lower::repr_inst_ident(n.as_str(), args)
-                    ));
-                }
-            }
-            let (n, is_rec) = named(&a[0])?;
-            let n_fn = crate::lower::drop_fn_ident(&n);
-            Some(if is_rec {
-                format!("__repr_list_rec_{n_fn}")
-            } else {
-                format!("__repr_list_{n_fn}")
-            })
-        }
-        Ty::Applied(TypeConstructorId::Option, a) if a.len() == 1 => {
-            // An instantiation payload (`${opt}` over `Option[Tree[String]]`).
-            if let Ty::Named(n, args) = &a[0] {
-                if !args.is_empty() && resolve_aggregate(&a[0], registry).is_none() {
-                    return Some(format!(
-                        "__repr_opt_{}",
-                        crate::lower::repr_inst_ident(n.as_str(), args)
-                    ));
-                }
-            }
-            let (n, is_rec) = named(&a[0])?;
-            let n_fn = crate::lower::drop_fn_ident(&n);
-            if !is_rec {
-                // A CUSTOM-variant payload (`${opt_tree}` over `Option[Tree]` — C-009):
-                // the generated `__repr_opt_<V>` (some/none over the variant repr).
-                // Option/Result payloads keep their existing table routing.
-                if matches!(&a[0], Ty::Named(vn, _)
-                    if !matches!(vn.as_str(), "Option" | "Result" | "Value"))
-                {
-                    return Some(format!("__repr_opt_{n_fn}"));
-                }
-                return None;
-            }
-            Some(format!("__repr_opt_rec_{n_fn}"))
-        }
-        // `${Map[String, <record/variant>]}` — the paired-slot map repr (quoted keys,
-        // element repr values, `[:]` when empty).
-        Ty::Applied(TypeConstructorId::Map, a)
-            if a.len() == 2 && matches!(a[0], Ty::String) =>
-        {
-            let (n, _is_rec) = named(&a[1])?;
-            let n_fn = crate::lower::drop_fn_ident(&n);
-            Some(format!("__repr_map_{n_fn}"))
-        }
-        _ => None,
-    }
-}
-
-/// Desugar a STRING INTERPOLATION `"…${e}…"` into a left-nested `ConcatStr` fold,
-/// seeded by an empty `""` literal: `(((("" ++ p0) ++ p1) … ) ++ p_{K-1})`. Each
-/// part is wrapped in its type's `to_string` ([`interp_part_leaf`]) — a Lit/String
-/// part is a no-call leaf, every other part a single `module.to_string` call.
-/// Concatenating with the leading `""` is byte-identical to v0's `emit_string_interp`
-/// (`"" ++ bytes == bytes`), so the folded String matches v0 in EVERY position.
-///
-/// This is the SINGLE source the lowering ([`LowerCtx::try_lower_string_interp`])
-/// AND the corpus caps gate (`count_ir_calls` in classify_corpus) BOTH consult: the
-/// gate counts the call NODES of the very tree the lowering emits, so the synthetic
-/// MIR `Op::CallFn`s are 1:1 backed by IR call nodes — `mir_calls == ir_calls` for an
-/// in-profile interp BY CONSTRUCTION (no `mir > ir` over-count, no spurious caps
-/// taint). Soundness rests on one invariant: when this returns `Some(tree)`, every
-/// leaf lowers to exactly one `CallFn` (a pure `module.to_string`, admitted by
-/// `purity::is_pure`) or a no-call passthrough — so `try_lower_concat_str` never
-/// rolls back. Returns `None` (the interp stays the deferred Opaque, credited 0 by
-/// the gate) iff a part has no admitted `to_string` module — a memory-safe defer.
-///
-/// THE WALL DOES THE HEAVY LIFTING: a part whose `to_string` is UNLINKED (Float /
-/// compound — registered in `PURE_MODULES` but not in the self-host runtime) still
-/// desugars to a real `CallFn`, so the enclosing function emits an unlinked call and
-/// the render wall (`try_render_wasm_program`) REJECTS it as `Unsupported`. Such a
-/// function is OUT of profile, so it can never contribute a `count != lower`
-/// mismatch — the only IN-profile interps are the fully-linkable ones (Lit/String/
-/// Int/Bool), where `count == lower` is trivially exact.
-pub fn desugar_string_interp(parts: &[IrStringPart], registry: &RecordLayouts) -> Option<IrExpr> {
-    let mut acc = lit_str("");
-    for p in parts {
-        let leaf = interp_part_leaf(p, registry)?;
-        acc = IrExpr {
-            kind: IrExprKind::BinOp {
-                op: almide_ir::BinOp::ConcatStr,
-                left: Box::new(acc),
-                right: Box::new(leaf),
-            },
-            ty: Ty::String,
-            span: None,
-            def_id: None,
-        };
-    }
-    Some(acc)
-}
-
-/// The SYNTHETIC call names the recursive Display ([`display_value`]) introduces for a
-/// single value of type `ty` — the `<module>.to_string`-family wrappers, recursively. A
-/// scalar/string/float/list value contributes ONE name; a record/tuple value contributes
-/// none itself but recurses via [`aggregate_synthetic_names`] into its fields. This DOES
-/// NOT count the value's OWN inner calls (it counts the WRAPPERS the desugar adds, not the
-/// operand) — keeping the `count_ir_calls` operand-descent free of double counting.
-fn value_synthetic_names(ty: &Ty, registry: &RecordLayouts, out: &mut Vec<String>) {
-    match ty {
-        // A nested record/tuple expands INLINE (recursive `__str_concat` + field formatters).
-        Ty::Record { .. } | Ty::Tuple(_) | Ty::Named(..) if resolve_aggregate(ty, registry).is_some() => {
-            aggregate_synthetic_names(ty, registry, out);
-        }
-        // Every OTHER value type routes to exactly ONE `to_string`-family call — the SAME single
-        // wrapper [`display_value`] / [`interp_part_leaf`] emit (Int → int.to_string, Float →
-        // float.to_string_compound, String → string.quote, List → list.to_string*, Map/Set/Option/
-        // Result → the unlinked `<module>.to_string` that walls). Keyed off `display_leaf_call` so
-        // the gate's count is BY CONSTRUCTION the lowering's emitted call set.
-        _ => {
-            if let Some((m, f)) = display_leaf_call(ty) {
-                out.push(format!("{m}.{f}"));
-            }
-        }
-    }
-}
-
-/// The SYNTHETIC call names the recursive Display ([`display_aggregate`]) introduces for an
-/// aggregate of type `ty`: one `__str_concat` per `ConcatStr` fold the expansion builds
-/// (= the number of `concat_all` parts at this level) plus the field formatters recursively.
-/// MIRRORS `display_aggregate`'s structure EXACTLY so the gate credits precisely the
-/// synthetic CallFns the lowering emits (count == lower for the aggregate, by construction).
-fn aggregate_synthetic_names(ty: &Ty, registry: &RecordLayouts, out: &mut Vec<String>) {
-    // A non-resolvable aggregate (structural record, unregistered) yields no Display tree —
-    // the part declines and the whole interp credits 0 (matched by `interp_synthetic_call_names`).
-    let Some((type_name, is_tuple, fields)) = resolve_aggregate(ty, registry) else {
-        return;
-    };
-    if !is_tuple && type_name.is_none() {
-        return; // structural record has no Display → walls, credits 0
-    }
-    // `concat_all` parts at this level: opening + (per field: a leading ", " for idx>0,
-    // a "field: " label for a record, the field formatter) + closing.
-    //   record: 1 (open) + Σ_i [ (i>0 → 1) + 1 (label) + 1 (formatter) ] + 1 (close)
-    //   tuple:  1 (open) + Σ_i [ (i>0 → 1) +            1 (formatter) ] + 1 (close)
-    let mut concat_parts = 2; // open + close
-    for (idx, _) in fields.iter().enumerate() {
-        if idx > 0 {
-            concat_parts += 1; // ", "
-        }
-        if !is_tuple {
-            concat_parts += 1; // "field: "
-        }
-        concat_parts += 1; // the field formatter expression
-    }
-    for _ in 0..concat_parts {
-        out.push("__str_concat".to_string());
-    }
-    for (_, fty) in &fields {
-        value_synthetic_names(fty, registry, out);
-    }
-}
-
-/// Count the synthetic `CallFn`s [`desugar_string_interp`] yields for `parts` — the
-/// `ConcatStr` and `module.to_string`-family call NODES of the desugared tree. The corpus
-/// gate adds exactly this to its IR call count for each interp (it counts the same tree),
-/// so the MIR calls the lowering emits are 1:1 backed. `None` (a part with no admitted
-/// Display) ⇒ 0 (the interp stays Opaque, lowering emits no synthetic call).
-pub fn interp_str_synthetic_call_count(parts: &[IrStringPart], registry: &RecordLayouts) -> usize {
-    interp_synthetic_call_names(parts, registry).len()
-}
-
-/// The SYNTHETIC call names [`desugar_string_interp`] introduces for `parts`: one
-/// `__str_concat` per TOP-LEVEL fold step (= `parts.len()`: K parts over the `""` seed ⇒ K
-/// concats) and, per non-passthrough part, the Display wrappers it adds — a scalar part one
-/// `<module>.to_string`, a RECORD/TUPLE part the full recursive `__str_concat` + field-
-/// formatter set ([`aggregate_synthetic_names`]). It DOES NOT include the operands' OWN
-/// inner calls (a `${g(x)}` callee) — those live in the original part exprs and are reached
-/// separately by `count_ir_calls`'s descent, so no double count. Empty (a `None` desugar —
-/// a part with no admitted Display) ⇒ the interp stays Opaque, crediting none.
-pub fn interp_synthetic_call_names(parts: &[IrStringPart], registry: &RecordLayouts) -> Vec<String> {
-    // A part with no admitted Display ⇒ the whole interp is non-desugarable (the lowering
-    // returns `None` and defers to Opaque), so it credits zero synthetic calls.
-    if desugar_string_interp(parts, registry).is_none() {
-        return Vec::new();
-    }
-    let mut names = Vec::with_capacity(parts.len() * 2);
-    // The TOP-LEVEL fold: K parts over the `""` seed ⇒ K `__str_concat` (the interp's own
-    // outer concatenation — a record/tuple part is ONE top-level part here, its INNER
-    // `__str_concat`s are added by `value_synthetic_names` below).
-    for _ in 0..parts.len() {
-        names.push("__str_concat".to_string());
-    }
-    for p in parts {
-        let IrStringPart::Expr { expr } = p else {
-            continue;
-        };
-        if matches!(expr.ty, Ty::String) {
-            continue; // a String part is a no-call passthrough
-        }
-        // A TOP-LEVEL record/tuple part mirrors `interp_part_leaf`'s decision tree
-        // EXACTLY (the mir == ir contract): an ANON record is ALWAYS one generated
-        // `__repr_anonrec_<hash>` call; an expand-foldable named/tuple part credits the
-        // full recursive tree; a non-expandable NAMED record one `__repr_rec_<R>`; any
-        // other non-expandable aggregate one `compound.to_string` (the wall).
-        if matches!(expr.ty, Ty::Record { .. } | Ty::Tuple(_) | Ty::Named(..))
-            && resolve_aggregate(&expr.ty, registry).is_some()
-        {
-            if let Ty::Record { fields } = &expr.ty {
-                names.push(format!(
-                    "__repr_{}",
-                    crate::lower::anon_record_drop_name(fields)
-                ));
-            } else if aggregate_part_expandable(expr, registry) {
-                aggregate_synthetic_names(&expr.ty, registry, &mut names);
-            } else if let Ty::Named(name, _) = &expr.ty {
-                names.push(format!(
-                    "__repr_rec_{}",
-                    crate::lower::drop_fn_ident(name.as_str())
-                ));
-            } else {
-                names.push("compound.to_string".to_string());
-            }
-        } else if let Some(n) = container_repr_name(&expr.ty, registry) {
-            // Mirrors `interp_part_leaf`'s container-repr arm: ONE generated call node.
-            names.push(n);
-        } else {
-            value_synthetic_names(&expr.ty, registry, &mut names);
-        }
-    }
-    names
-}
-
-/// Is a WHOLE interpolation DESUGARABLE (every part has an admitted Display)? When true, the
-/// lowering folds it to a `ConcatStr` chain; when false, it stays the deferred Opaque.
-/// (Desugarable does NOT imply LINKABLE — a Float part desugars but float.to_string is
-/// unlinked, so the function walls at render. Use the registry to split proven-vs-walled;
-/// this predicate only answers "does the lowering fold it".)
-pub fn interp_str_desugarable(parts: &[IrStringPart], registry: &RecordLayouts) -> bool {
-    desugar_string_interp(parts, registry).is_some()
-}
-
-/// Does `module.func` return a real MATERIALIZED `Result[Int, String]` (the DynListStr len-as-tag
-/// layout)? Its result may be tracked in `materialized_results` so an `Ok`/`Err` `match` over it
-/// EXECUTES. NARROW to fns actually self-hosted — any other Result is a deferred `Opaque` (len 0,
-/// would misread as `Ok`). `int.parse` is the canonical for string.to_int/to_integer/parse_int.
-/// The CallFn name for a stdlib `module.func` call, routing the REPR-POLYMORPHIC list combinators
-/// to their `_str` variant when the RESULT is a `List[heap]` (e.g. `list.map` over a `List[String]`
-/// → `list.map_str`, a DynListStr-result impl). The element repr (i64 vs i32 handle) demands a
-/// separate variant; the variant reads/writes via the heap-aware prim ops. Scalar-result lists keep
-/// the plain name. `module.func` is unchanged for everything else.
-pub(crate) fn list_heap_call_name(
-    module: &str,
-    func: &str,
-    arg_tys: &[Ty],
-    result_ty: &Ty,
-    // Is the Map KEY type (of the first-arg/result Map) a NULLARY-ONLY variant?
-    // Computed by the caller (LowerCtx has the variant_layouts; this router is a
-    // free fn) — gates the `_vtag` tag-normalized map family. `map_key_scalar_rec`
-    // is the all-Int/Bool-field record-key twin, gating `_srec`.
-    map_key_nullary: bool,
-    map_key_scalar_rec: bool,
-) -> String {
-    // A MONO-SPECIALIZED stdlib call name (`result.or_else__Int_String_String` —
-    // the optimizer suffixes a generic intrinsic's instantiation) must route by
-    // its BASE name: the registry links base names only, so the suffixed form
-    // fell through every router arm to an UNLINKED dotted name and walled the fn
-    // (fuzz B-198's or_else). The instantiation's types are already in
-    // `arg_tys`/`result_ty` — the suffix carries no information the router needs.
-    let func = func.split_once("__").map_or(func, |(base, _)| base);
-    // #781: the monolithic 780-line dispatch (cog 324) is decomposed into
-    // per-module routers. Routing ORDER is load-bearing and preserved: the
-    // heap-accumulator `fold` guard fires BEFORE the per-module tables (a
-    // scalar-acc fold over heap elements falls through to `list.fold_str`).
-    if module == "random" && matches!(func, "choice" | "shuffle") {
-        return random_call_name(func, arg_tys);
-    }
-    if module == "fan" && func == "map" {
-        return fan_map_call_name(arg_tys, result_ty);
-    }
-    if func == "fold" && matches!(module, "list" | "map" | "set") && is_heap_ty(result_ty) {
-        return heap_fold_call_name(module, arg_tys, result_ty);
-    }
-    let routed = match module {
-        "list" => list_call_name(func, arg_tys, result_ty),
-        "set" => set_call_name(func, arg_tys, result_ty),
-        "map" => map_call_name(func, arg_tys, result_ty, map_key_nullary, map_key_scalar_rec),
-        "result" | "option" if func == "unwrap_or" => unwrap_or_call_name(module, arg_tys),
-        "option" => option_call_name(func, arg_tys, result_ty),
-        "result" => result_call_name(func, arg_tys, result_ty),
-        // `value.keys` IS `json.keys` (one impl, two stdlib names) — remap to the
-        // registered self-host; every other value.* rides its own dotted name.
-        "value" if func == "keys" => Some("json.keys".to_string()),
-        _ => None,
-    };
-    routed.unwrap_or_else(|| format!("{module}.{func}"))
-}
-
-/// Route the payload-polymorphic `option` combinators by PAYLOAD/RESULT repr. The
-/// self-host impls (`option_map.almd`) are `Option[Int]`-typed — SCALAR payloads
-/// only (Int/Bool/Float ride the i64 slot identically). A HEAP payload or result
-/// (`option.map(some("hi"), (s) => s + "!")`) invoked the scalar impl anyway: the
-/// closure declares the `$closure_fnN_h` (i32-result) type while `__opt_map_some`
-/// calls through `$closure_fnN` (i64) — the "indirect call type mismatch" TRAP on
-/// the verified default (the #790 option.map row, main-reachable). Route those to
-/// an UNREGISTERED wall suffix instead — the fn walls, v0 runs the shape correctly
-/// (the same honest-wall pattern as the map `_skv_wall` family).
-fn option_call_name(func: &str, arg_tys: &[Ty], result_ty: &Ty) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId as TC;
-    // `option.to_list` keys on the PAYLOAD: a flat heap payload (String /
-    // List[scalar] / scalar tuple) rides the co-owning `_rc` variant (the raw slot
-    // copy aliased the payload un-owned — double free); a richer payload walls.
-    if func == "to_list" {
-        if let Some(Ty::Applied(TC::Option, a)) = arg_tys.first() {
-            if a.len() == 1 && is_heap_ty(&a[0]) {
-                if matches!(a[0], Ty::String) || is_flat_scalar_block_ty(&a[0]) {
-                    return Some("option.to_list_rc".to_string());
-                }
-                return Some("option.to_list_x".to_string());
-            }
-        }
-    }
-    // ONE mismatch axis is the CLOSURE's RESULT repr: params always ride the
-    // widened i64 slots, and an Option-returning closure uses the same `_h` table
-    // type the impl declares (flat_map / or_else match by construction; filter's
-    // pred is scalar-result; flatten / zip take no closure at all). The two shapes
-    // whose USER closure result repr can diverge from the scalar-typed impl:
-    //   - `option.map` with a HEAP mapped payload (impl `f: (Int) -> Int` = i64
-    //     result; a `(s) => s + "!"` closure declares the i32 `_h` type)
-    //   - `option.unwrap_or_else` with a HEAP payload (impl `f: () -> Int`)
-    let heap_option =
-        |t: &Ty| matches!(t, Ty::Applied(TC::Option, a) if a.len() == 1 && is_heap_ty(&a[0]));
-    match func {
-        // The heap twins declare the closure heap-typed, so the `_h` CallIndirect
-        // table type matches by construction (option_map.almd's `_h` family).
-        "map" if heap_option(result_ty) => Some("option.map_h".to_string()),
-        // filter/flatten/or_else heap twins: the OTHER axis is OWNERSHIP — the
-        // kept payload must SHARE (Dup) into the rebuilt some(); the scalar
-        // rewrap raw-copied the handle un-owned (or_else: fuzz seed-20260718
-        // index 622, correct output then an __rc_dec trap at scope end).
-        "filter" if heap_option(result_ty) => Some("option.filter_h".to_string()),
-        "flatten" if heap_option(result_ty) => Some("option.flatten_h".to_string()),
-        "or_else" if heap_option(result_ty) => Some("option.or_else_h".to_string()),
-        // `to_result` over ANY heap payload: the heap twin builds the CAP-AS-TAG
-        // Result the consumers read (the scalar impl's len-as-tag misread). The twin's
-        // internals are payload-type-independent (one handle slot, Dup-shared into
-        // ok()), so a NESTED heap payload (`Option[Result[Int, String]]` —
-        // to_result_nested_share) rides the same routine as the String payload; the
-        // co-own (+1) discipline is exactly the fix for the v0 double-free that
-        // fixture pinned.
-        "to_result"
-            if matches!(arg_tys.first(), Some(Ty::Applied(TC::Option, a))
-                if a.len() == 1 && is_heap_ty(&a[0])) =>
-        {
-            Some("option.to_result_h".to_string())
-        }
-        "unwrap_or_else" if is_heap_ty(result_ty) => {
-            Some("option.unwrap_or_else_h".to_string())
-        }
-        _ => None,
-    }
-}
-
-/// Route `result.partition` / `result.collect_map` by REPR: the self-host impls
-/// (result_collect.almd) pin the Ok payload to a SCALAR raw-copy slot and the Err
-/// to a deep-copied String. Any other repr routes to the UNLINKED `_x` suffix — a
-/// clean render wall, never a wrong-typed link (the random/fan `_x` discipline).
-/// The mapped-list ELEMENT (collect_map's `xs`) is unconstrained: its slot is
-/// forwarded to the closure untouched (borrowed, any repr).
-fn result_call_name(func: &str, arg_tys: &[Ty], result_ty: &Ty) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId as TC;
-    match func {
-        // is_ok/is_err over a HEAP-Ok Result: the block is cap-as-tag (tag @16, len@4
-        // always 1) — the len-as-tag base impls would call every heap-Ok value an Err.
-        "is_ok" | "is_err"
-            if matches!(arg_tys.first(), Some(Ty::Applied(TC::Result, a))
-                if a.len() == 2 && is_heap_ty(&a[0])) =>
-        {
-            Some(format!("result.{func}_h"))
-        }
-        // The VALUE combinators over a HEAP-Ok Result — same cap-as-tag misread as
-        // is_ok/is_err, but the scalar impls also REBUILT the wrong layout: every
-        // `ok(x)` took the Err path and the result printed as a swapped/zeroed value
-        // (the fuzz C-904 silent `ok("")` class; unwrap_or_else even emitted invalid
-        // wasm — an i64-result CallFn bound to an i32 String local). The exact
-        // `Result[String, String]` instantiation routes to the `_h` twins
-        // (result_map.almd); any other heap-Ok instantiation routes to the UNLINKED
-        // `_x` — a deterministic render wall, never a wrong-typed link. `unwrap_or`
-        // needs no arm (unwrap_or_call_name already keys on the repr), and the
-        // display/eq families are chosen by type at the call site.
-        "map" | "map_err" | "flat_map" | "unwrap_or_else" | "to_option"
-        | "to_err_option" | "filter" | "or_else" | "flatten" | "to_list" | "zip"
-            if matches!(arg_tys.first(), Some(Ty::Applied(TC::Result, a))
-                if a.len() == 2 && is_heap_ty(&a[0])) =>
-        {
-            // `flatten`'s BASE impl already reads the heap-Ok OUTER (tag@16) and the
-            // len-as-tag scalar INNER — `Result[Result[scalar, String], String]` is
-            // its exact shape (option_result_symmetry); only other inners wall.
-            if func == "flatten" {
-                let base_ok = matches!(arg_tys.first(), Some(Ty::Applied(TC::Result, a))
-                    if matches!(&a[0], Ty::Applied(TC::Result, i)
-                        if i.len() == 2 && !is_heap_ty(&i[0]) && matches!(i[1], Ty::String))
-                        && matches!(a[1], Ty::String));
-                return Some(if base_ok {
-                    "result.flatten".to_string()
-                } else {
-                    "result.flatten_x".to_string()
-                });
-            }
-            let ss_in = matches!(arg_tys.first(), Some(Ty::Applied(TC::Result, a))
-                if matches!(a[0], Ty::String) && matches!(a[1], Ty::String));
-            let ss_res = matches!(result_ty, Ty::Applied(TC::Result, a)
-                if a.len() == 2 && matches!(a[0], Ty::String) && matches!(a[1], Ty::String));
-            let has_h = match func {
-                "map" | "map_err" | "flat_map" => ss_in && ss_res,
-                // The `_h` twin is payload-type-INDEPENDENT (one handle slot per side:
-                // the Ok returns with the share discipline, the Err borrows into f, the
-                // closure ABI is (i32) -> i32 for any heap in/out), so ANY heap-Ok /
-                // heap-Err instantiation admits — `Result[List[Float], List[String]]`
-                // (uoe_heap_ok_share via result.collect) rides the String routine. A
-                // SCALAR Err stays walled: its slot holds an i64 value, not a handle,
-                // and the twin's load_handle would misread it.
-                "unwrap_or_else" => {
-                    matches!(arg_tys.first(), Some(Ty::Applied(TC::Result, a))
-                        if a.len() == 2 && is_heap_ty(&a[1]))
-                        && is_heap_ty(result_ty)
-                }
-                // `to_option`'s `_h` twin is payload-type-INDEPENDENT (Ok → the
-                // shared handle into some(), Err → none — no Err read at all), so
-                // ANY heap-Ok instantiation admits (`result.to_option(v3)` over
-                // `Result[Option[Float], String]` — C-149).
-                "to_option" => true,
-                "to_err_option" => ss_in,
-                _ => false,
-            };
-            Some(if has_h { format!("result.{func}_h") } else { format!("result.{func}_x") })
-        }
-        // The same combinators with a SCALAR-Ok INPUT but a HEAP-Ok RESULT
-        // (`result.map(r, (v) => some(v))` — `Result[Int, String]` →
-        // `Result[Option[Int], String]`, fuzz seed-20260718 index 647): the
-        // scalar impl's i64-result closure table type mismatches the
-        // heap-result closure AND its len-as-tag rebuild is the wrong OUTPUT
-        // layout. `map` routes to the `_s2h` twin (C-151 — Value-erased heap
-        // payload, cap-as-tag build); map_err/flat_map keep the UNLINKED `_x`.
-        // (Reached only when the input arm above did not match.)
-        "map" | "map_err" | "flat_map"
-            if matches!(result_ty, Ty::Applied(TC::Result, a)
-                if a.len() == 2 && is_heap_ty(&a[0])) =>
-        {
-            let err_is_string = matches!(result_ty, Ty::Applied(TC::Result, a)
-                if a.len() == 2 && matches!(a[1], Ty::String));
-            if func == "map" && err_is_string {
-                return Some("result.map_s2h".to_string());
-            }
-            Some(format!("result.{func}_x"))
-        }
-        // partition: List[Result[scalar, String]] → (List[scalar], List[String])
-        "partition" => {
-            let ok = matches!(arg_tys.first(), Some(Ty::Applied(TC::List, e))
-                if e.len() == 1
-                    && matches!(&e[0], Ty::Applied(TC::Result, re)
-                        if re.len() == 2 && !is_heap_ty(&re[0]) && matches!(re[1], Ty::String)));
-            Some(if ok { "result.partition".to_string() } else { "result.partition_x".to_string() })
-        }
-        // collect_map: (List[T], (T) -> Result[U, E]) → Result[List[U], List[E]]
-        // with U scalar, E String (keyed on the RESULT type — the closure's own
-        // repr is heap-result `_h` by construction for every instantiation).
-        "collect_map" => {
-            let ok = matches!(result_ty, Ty::Applied(TC::Result, oe)
-                if oe.len() == 2
-                    && matches!(&oe[0], Ty::Applied(TC::List, u) if u.len() == 1 && !is_heap_ty(&u[0]))
-                    && matches!(&oe[1], Ty::Applied(TC::List, e) if e.len() == 1 && matches!(e[0], Ty::String)));
-            Some(if ok {
-                "result.collect_map".to_string()
-            } else {
-                "result.collect_map_x".to_string()
-            })
-        }
-        _ => None,
-    }
-}
-
-/// `random.choice`/`random.shuffle` routing (extracted from the monolithic
-/// dispatch — #781). Always names a target: the `_x` tail is the honest wall.
-fn random_call_name(func: &str, arg_tys: &[Ty]) -> String {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // `random.choice` / `random.shuffle` key on the LIST ELEMENT: a scalar element uses the
-    // flat self-host (i64 slots, flat drops), a String element the rc-aware `_str` variant
-    // (random_choice.almd / random_shuffle.almd). Any other element class routes to the
-    // UNLINKED `random.<func>_x` → a clean render wall (never a wrong-typed link).
-    if let Some(Ty::Applied(TypeConstructorId::List, a)) = arg_tys.first() {
-        if a.len() == 1 && !is_heap_ty(&a[0]) {
-            return format!("random.{func}");
-        }
-        if a.len() == 1 && matches!(a[0], Ty::String) {
-            return format!("random.{func}_str");
-        }
-    }
-    format!("random.{func}_x")
-}
-
-/// `fan.map` monomorphic-variant routing (see the block comment).
-fn fan_map_call_name(arg_tys: &[Ty], result_ty: &Ty) -> String {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // `fan.map` selects a monomorphic self-host by (input element A, output element B) — `fan.map<sfx>`
-    // where A/B in {Int (""), String ("s")}. The input A is `arg_tys[0] = List[A]`; the output B is
-    // `result_ty = Result[List[B], String]`. An unsupported pairing routes to the UNLINKED
-    // `fan.map_x` → a clean render wall (never a wrong-typed link = never invalid wasm).
-    fn elem_of(t: Option<&Ty>) -> Option<&Ty> {
-        match t {
-            Some(Ty::Applied(TypeConstructorId::List, e)) if e.len() == 1 => Some(&e[0]),
-            _ => None,
-        }
-    }
-    let a = elem_of(arg_tys.first());
-    let b = match result_ty {
-        Ty::Applied(TypeConstructorId::Result, r) if r.len() == 2 => elem_of(Some(&r[0])),
-        _ => None,
-    };
-    let sfx = match (a, b) {
-        (Some(Ty::Int), Some(Ty::Int)) => "",
-        (Some(Ty::Int), Some(Ty::String)) => "_is",
-        (Some(Ty::String), Some(Ty::String)) => "_ss",
-        (Some(Ty::String), Some(Ty::Int)) => "_si",
-        _ => "_x",
-    };
-    format!("fan.map{sfx}")
-}
-
-/// HEAP-accumulator `fold` routing for list/map/set (see the block comment).
-fn heap_fold_call_name(module: &str, arg_tys: &[Ty], result_ty: &Ty) -> String {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // `fold` threads an ACCUMULATOR (= the result type). A HEAP accumulator (e.g. a String built up
-    // across the fold) needs the closure-result + accumulator to be an i32 handle, not the i64 the
-    // scalar-accumulator fold variants hardcode — emitting an i32 there is invalid wasm. No heap-
-    // accumulator fold variant is self-hosted yet, so route it to an UNREGISTERED name: render walls
-    // it cleanly (a controlled reject) rather than emitting a repr-mismatched module. (Soundness-
-    // preserving: a wall is never a miscompile.)
-    // The ONE self-hosted heap-accumulator variant: `Option[List[String]]` acc over
-    // `List[String]` elements (is_balanced's paren-stack fold) — `list.fold_ols`, the
-    // (heap, heap) -> heap 2-arity CallIndirect (`list_reduce_str`'s proven closure
-    // shape). Every other heap-acc fold keeps the unregistered `fold_hacc` wall.
-    let is_ols_acc = matches!(result_ty,
-        Ty::Applied(TypeConstructorId::Option, a) if a.len() == 1
-            && matches!(&a[0], Ty::Applied(TypeConstructorId::List, b)
-                if b.len() == 1 && matches!(b[0], Ty::String)));
-    let src_is_list_str = matches!(arg_tys.first(),
-        Some(Ty::Applied(TypeConstructorId::List, e)) if e.len() == 1 && matches!(e[0], Ty::String));
-    if module == "list" && is_ols_acc && src_is_list_str {
-        return "list.fold_ols".to_string();
-    }
-    // The `Map[String, Int]`-accumulator map.fold pair (the map_fold_heap_acc corpus
-    // shape): keyed on the EXACT acc type + subject family — `Map[String, String]`
-    // (fold_str_msi) / `Map[String, Int]` (fold_skv_msi). Other heap accs keep the wall.
-    let is_msi_acc = matches!(result_ty,
-        Ty::Applied(TypeConstructorId::Map, a) if a.len() == 2
-            && matches!(a[0], Ty::String) && matches!(a[1], Ty::Int));
-    // `list.fold` over a Map[String, Int] acc with String elements — the frequencies
-    // shape (map_fold_hacc.almd's list_fold_str_msi; same closure ABI as fold_ols).
-    if module == "list" && is_msi_acc && src_is_list_str {
-        return "list.fold_str_msi".to_string();
-    }
-    if module == "map" && is_msi_acc {
-        if let Some(Ty::Applied(TypeConstructorId::Map, s)) = arg_tys.first() {
-            if s.len() == 2 && matches!(s[0], Ty::String) {
-                if matches!(s[1], Ty::String) {
-                    return "map.fold_str_msi".to_string();
-                }
-                if matches!(s[1], Ty::Int) {
-                    return "map.fold_skv_msi".to_string();
-                }
-            }
-        }
-    }
-    // A STRING accumulator over a `Map[String, String]` subject (map_fold.almd's
-    // `map.fold(ms, "[", (acc, k, v) => acc + k + "=" + v)` builder): all sides are
-    // handles, so the all-i32 3-arity CallIndirect twin runs it faithfully — the same
-    // interleaved-pair walk as fold_str_msi with a String acc.
-    if module == "map" && matches!(result_ty, Ty::String) {
-        if let Some(Ty::Applied(TypeConstructorId::Map, s)) = arg_tys.first() {
-            if s.len() == 2 && matches!(s[0], Ty::String) && matches!(s[1], Ty::String) {
-                return "map.fold_str_sacc".to_string();
-            }
-        }
-    }
-    // A RECORD/VARIANT accumulator over a `List[record/variant]` (`list.fold(counters,
-    // Counter.empty(), (acc, c) => Counter.merge(acc, c))` — the protocol-merge shape,
-    // and the cross-module `Loc` pick fold): every side is a uniform heap HANDLE, so the
-    // TYPE-ERASED `list.fold_hrec` (the fold_ols closure ABI) runs it faithfully — acc
-    // MOVES into f each step, elements are borrowed, the final acc is the owned result.
-    // A Named alias to a scalar fails `is_heap_ty` and keeps the wall.
-    let is_named_heap =
-        |t: &Ty| matches!(t, Ty::Named(..) | Ty::Record { .. }) && is_heap_ty(t);
-    let elem_is_named = matches!(arg_tys.first(),
-        Some(Ty::Applied(TypeConstructorId::List, e)) if e.len() == 1 && is_named_heap(&e[0]));
-    if module == "list" && elem_is_named && is_named_heap(result_ty) {
-        return "list.fold_hrec".to_string();
-    }
-    // A heap accumulator over SCALAR Int/Bool elements (`list.fold([true, false],
-    // tmp, (s, v) => ok(…))` with a Result acc — result_uoe_float): the elements are
-    // plain i64 slot copies, the acc a handle, so the (heap, scalar) -> heap
-    // 2-arity CallIndirect twin runs it faithfully. Float elements stay walled (an
-    // f64 closure param would ABI-mismatch the i64 slot load).
-    let elem_is_int_bool = matches!(arg_tys.first(),
-        Some(Ty::Applied(TypeConstructorId::List, e)) if e.len() == 1 && matches!(e[0], Ty::Int | Ty::Bool));
-    if module == "list" && elem_is_int_bool && is_heap_ty(result_ty) {
-        return "list.fold_hsca".to_string();
-    }
-    format!("{module}.fold_hacc")
-}
-
-/// Repr-poly `result.unwrap_or` / `option.unwrap_or` routing (see the block
-/// comment). Caller guarantees `func == "unwrap_or"`.
-fn unwrap_or_call_name(module: &str, arg_tys: &[Ty]) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // `option.unwrap_or(o, d)` over an `Option[String]` (the pipe/UFCS form
-    // `list.get(xs, i) |> option.unwrap_or("")`, NOT the `??` operator that
-    // `try_lower_option_unwrap_or` desugars) must route to `option.unwrap_or_str`: the
-    // generic `option.unwrap_or` takes its default as an i64 SCALAR (`Option[Int]`), so a
-    // String fallback (an i32 handle) and a String result repr-mismatch it — invalid wasm
-    // (`expected i64, found i32` in the call + the i64 result). `option.unwrap_or_str`
-    // (param i32 i32) (result i32) is the rc-correct String variant (deep-copies the kept
-    // payload so result + source can both drop). Keyed on the Option payload being String.
-    // The same repr-poly routing for `result.unwrap_or` (the pipe/direct form —
-    // `result.unwrap_or(json.parse(s), json.null())`, json_gltf_walk): the generic
-    // impl takes an i64 scalar default; a heap Ok/default needs the rc-correct
-    // self-hosts registered for the `??` desugar.
-    if module == "result" {
-        if let Some(Ty::Applied(TypeConstructorId::Result, a)) = arg_tys.first() {
-            if a.len() == 2 && is_value_ty(&a[0]) {
-                return Some("result.value_unwrap_or".to_string());
-            }
-            if a.len() == 2
-                && matches!(&a[0], Ty::Applied(TypeConstructorId::List, e)
-                    if e.len() == 1 && is_value_ty(&e[0]))
-            {
-                return Some("result.list_value_unwrap_or".to_string());
-            }
-            if a.len() == 2 && matches!(a[0], Ty::String) {
-                return Some("result.str_unwrap_or".to_string());
-            }
-            // A FLAT scalar block payload — a scalar TUPLE (`result.zip`'s `(Int, Int)`),
-            // a List[<scalar>], Bytes, or an `Option[<scalar>]` (the C-149
-            // nested-share chain — len-as-tag + one scalar slot, flat rc_dec):
-            // the rc-correct flat variant over the cap-as-tag layout (tag @16).
-            if a.len() == 2
-                && (matches!(&a[0], Ty::Tuple(ts) if !ts.is_empty() && ts.iter().all(|t| !is_heap_ty(t)))
-                    || matches!(&a[0], Ty::Applied(TypeConstructorId::List, e)
-                        if e.len() == 1 && !is_heap_ty(&e[0]))
-                    || matches!(&a[0], Ty::Applied(TypeConstructorId::Option, e)
-                        if e.len() == 1 && !is_heap_ty(&e[0]))
-                    || matches!(a[0], Ty::Bytes))
-            {
-                return Some("result.flat_unwrap_or".to_string());
-            }
-            // Any OTHER heap Ok payload has no registered rc-correct variant: the
-            // generic impl takes an i64 SCALAR default, so a handle payload/default
-            // repr-mismatches (invalid wasm). Route to an unregistered `_x` name —
-            // the caller WALLS honestly.
-            if a.len() == 2 && is_heap_ty(&a[0]) {
-                return Some("result.unwrap_or_hx".to_string());
-            }
-        }
-    }
-    if module == "option" {
-        if let Some(Ty::Applied(TypeConstructorId::Option, a)) = arg_tys.first() {
-            if a.len() == 1 && matches!(a[0], Ty::String) {
-                return Some("option.unwrap_or_str".to_string());
-            }
-            // The remaining heap payloads route to their rc-correct self-hosts
-            // (already registered for the `??` desugar): Value / List[Value] /
-            // List[String]. The generic unwrap_or takes an i64 scalar default,
-            // so a handle default (`json.get_array(v,k) |> option.unwrap_or([])`,
-            // json_gltf_walk's count_floats) repr-mismatched — invalid wasm.
-            if a.len() == 1 && is_value_ty(&a[0]) {
-                return Some("option.value_unwrap_or".to_string());
-            }
-            if a.len() == 1
-                && matches!(&a[0], Ty::Applied(TypeConstructorId::List, e)
-                    if e.len() == 1 && is_value_ty(&e[0]))
-            {
-                return Some("option.listvalue_unwrap_or".to_string());
-            }
-            if a.len() == 1
-                && matches!(&a[0], Ty::Applied(TypeConstructorId::List, e)
-                    if e.len() == 1 && matches!(e[0], Ty::String))
-            {
-                return Some("option.liststr_unwrap_or".to_string());
-            }
-            // A FLAT scalar-element list payload (`map.get(groups, "0") ?? []` —
-            // Option[List[Int]], the group_by class): the rc-correct flat variant.
-            // A scalar TUPLE or Bytes payload is the SAME flat block shape (uniform
-            // slots / raw bytes @12, flat rc_dec drop) — bytes.chunks' element class.
-            if a.len() == 1
-                && (matches!(&a[0], Ty::Applied(TypeConstructorId::List, e)
-                        if e.len() == 1 && !is_heap_ty(&e[0]))
-                    || matches!(&a[0], Ty::Tuple(ts) if !ts.is_empty() && ts.iter().all(|t| !is_heap_ty(t)))
-                    // A NESTED `Option[<scalar>]` payload (`option.unwrap_or(
-                    // result.to_option(v3), none)` — C-149) is the SAME flat block
-                    // (len-as-tag + one scalar slot, flat rc_dec).
-                    || matches!(&a[0], Ty::Applied(TypeConstructorId::Option, e)
-                        if e.len() == 1 && !is_heap_ty(&e[0]))
-                    || matches!(a[0], Ty::Bytes))
-            {
-                return Some("option.listint_unwrap_or".to_string());
-            }
-            // Any OTHER heap payload: no registered rc-correct variant — wall honestly
-            // (the generic impl's i64 scalar default repr-mismatches a handle).
-            if a.len() == 1 && is_heap_ty(&a[0]) {
-                return Some("option.unwrap_or_hx".to_string());
-            }
-        }
-    }
-    None
-}
-
-/// Every `list.*` typed-variant route (extracted verbatim, order preserved;
-/// `None` = no typed variant applies → the caller falls to the plain name).
-fn list_call_name(func: &str, arg_tys: &[Ty], result_ty: &Ty) -> Option<String> {
-    // Split (2026-07-20, #781 cog>100 burn-down) into per-theme routers, called in the
-    // SAME ORDER the original single if-chain evaluated them (order is load-bearing: e.g.
-    // `enumerate` has two historical arms and the FIRST one — source_keyed — is exhaustive
-    // for any single-type-arg List, making transform's `enumerate` arm dead by construction,
-    // exactly as in the original unified function). Pure text move, no logic change.
-    list_call_name_hof_combinators(func, arg_tys, result_ty)
-        .or_else(|| list_call_name_source_keyed(func, arg_tys, result_ty))
-        .or_else(|| list_call_name_ordering(func, arg_tys, result_ty))
-        .or_else(|| list_call_name_transform(func, arg_tys, result_ty))
-        .or_else(|| list_call_name_modifiers(func, arg_tys, result_ty))
-        .or_else(|| list_call_name_accessors(func, arg_tys, result_ty))
-}
-
-fn list_call_name_hof_combinators(func: &str, arg_tys: &[Ty], result_ty: &Ty) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // `list.shuffle` / `list.choice` ARE `random.shuffle`/`random.choice` under a second
-    // stdlib name (the same almide_rt intrinsics) — delegate to the random element-repr
-    // router. The Entropy capability stays honest: the witness derives from the LINKED
-    // self-host body (prim.random_get), not the call-site module name.
-    if matches!(func, "shuffle" | "choice") {
-        return Some(random_call_name(func, arg_tys));
-    }
-    // `list.group_by` — the hval-map builder (scalar elements, String keys). Any other
-    // repr routes to the UNLINKED `_x` (a clean render wall, never a wrong-typed link).
-    if func == "group_by" {
-        let ok = matches!(arg_tys.first(), Some(Ty::Applied(TypeConstructorId::List, e))
-                if e.len() == 1 && !is_heap_ty(&e[0]))
-            && matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                if a.len() == 2
-                    && matches!(a[0], Ty::String)
-                    && matches!(&a[1], Ty::Applied(TypeConstructorId::List, b)
-                        if b.len() == 1 && !is_heap_ty(&b[0])));
-        return Some(if ok { "list.group_by".to_string() } else { "list.group_by_x".to_string() });
-    }
-    // `list.zip_with` keys on the RESULT element (= the closure's result repr, the
-    // only axis of the CallIndirect table type — params ride the widened i64 slots
-    // uniformly): a SCALAR result element rides the base impl (heap SOURCE elements
-    // are passed as borrowed handles, never copied into the flat result — sound);
-    // a (String, String → String) triple routes to the `_str` twin (move-in fill);
-    // any other heap result element routes to the UNLINKED `_x` — the scalar impl's
-    // $closure_fn2 table type TRAPS on a heap-result closure ("indirect call type
-    // mismatch", fuzz G-65) and its raw copies would alias handles un-owned.
-    if func == "zip_with" {
-        fn elem(t: Option<&Ty>) -> Option<&Ty> {
-            use almide_lang::types::constructor::TypeConstructorId;
-            match t {
-                Some(Ty::Applied(TypeConstructorId::List, e)) if e.len() == 1 => Some(&e[0]),
-                _ => None,
-            }
-        }
-        let a = elem(arg_tys.first());
-        let b = elem(arg_tys.get(1));
-        let c = elem(Some(result_ty));
-        return Some(match (a, b, c) {
-            (Some(_), Some(_), Some(z)) if !is_heap_ty(z) => "list.zip_with".to_string(),
-            (Some(Ty::String), Some(Ty::String), Some(Ty::String)) => {
-                "list.zip_with_str".to_string()
-            }
-            _ => "list.zip_with_x".to_string(),
-        });
-    }
-    // `list.scan` keys on (element, ACC) reprs — the ACC is the closure's result
-    // and the OUTPUT element (one table-type axis, one layout axis): scalar/scalar
-    // rides the base impl, scalar/String the `_str` twin (move-in fill, borrow-back
-    // threading), anything else the UNLINKED `_x` wall — the scalar impl's i64 init
-    // param failed validation on a String acc ("expected i64, found i32", the v1
-    // edition of fuzz seed-20260718 index 259).
-    if func == "scan" {
-        let elem_scalar = matches!(arg_tys.first(), Some(Ty::Applied(TypeConstructorId::List, e))
-            if e.len() == 1 && !is_heap_ty(&e[0]));
-        return Some(match (elem_scalar, arg_tys.get(1)) {
-            (true, Some(a)) if !is_heap_ty(a) => "list.scan".to_string(),
-            (true, Some(Ty::String)) => "list.scan_str".to_string(),
-            _ => "list.scan_x".to_string(),
-        });
-    }
-    // `list.unique_by` keys on (element, KEY) reprs — the KEY is the closure's
-    // result (the CallIndirect table-type axis): scalar/scalar rides the base
-    // impl, scalar/String the `_sk` twin (content equality via string.eq), and
-    // anything else the UNLINKED `_x` wall — the scalar `(Int) -> Int` table
-    // type TRAPPED on a String-key closure ("indirect call type mismatch",
-    // fuzz seed-20260718 index 9, the unique_by edition of the zip_with class).
-    if func == "unique_by" {
-        let elem_scalar = matches!(arg_tys.first(), Some(Ty::Applied(TypeConstructorId::List, e))
-            if e.len() == 1 && !is_heap_ty(&e[0]));
-        let key_ty = match arg_tys.get(1) {
-            Some(Ty::Fn { ret, .. }) => Some(ret.as_ref()),
-            _ => None,
-        };
-        return Some(match (elem_scalar, key_ty) {
-            (true, Some(k)) if !is_heap_ty(k) => "list.unique_by".to_string(),
-            (true, Some(Ty::String)) => "list.unique_by_sk".to_string(),
-            _ => "list.unique_by_x".to_string(),
-        });
-    }
-    None
-}
-
-fn list_call_name_source_keyed(func: &str, arg_tys: &[Ty], result_ty: &Ty) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // `list.enumerate` keys on its SOURCE element: scalar → the flat-pair self-host;
-    // String → the rc-share pair variant (`DropListIntStr` at the call site frees each
-    // pair's key ref); any other heap element routes to an UNREGISTERED name (walls
-    // cleanly — a flat pair drop would leak a rich element's children).
-    // `list.drop_end` keys on its SOURCE element: a String element routes to the CO-OWNED
-    // rc-copy variant (`__copy_slots_rc`) — the raw slot copy aliases heap handles un-owned,
-    // which double-frees under the nested Option[List[String]] drop (is_balanced's fold).
-    if func == "drop_end" {
-        if let Some(Ty::Applied(TypeConstructorId::List, a)) = arg_tys.first() {
-            if a.len() == 1 && matches!(a[0], Ty::String) {
-                return Some("list.drop_end_str".to_string());
-            }
-        }
-    }
-    if func == "enumerate" {
-        if let Some(Ty::Applied(TypeConstructorId::List, a)) = arg_tys.first() {
-            if a.len() == 1 {
-                if !is_heap_ty(&a[0]) {
-                    return Some("list.enumerate".to_string());
-                }
-                if matches!(a[0], Ty::String) {
-                    return Some("list.enumerate_str".to_string());
-                }
-                return Some("list.enumerate_h".to_string());
-            }
-        }
-    }
-    // `list.pop` keys on its SOURCE element: a SCALAR (8-byte-slot) element rides the
-    // registered in-place self-host (`list_pop.almd` — Int/Bool/Float move bit-exactly);
-    // a HEAP element routes to the UNREGISTERED `_x` name and walls at render (popping
-    // an owned handle is an ownership transfer the flat impl cannot express).
-    if func == "pop" {
-        if let Some(Ty::Applied(TypeConstructorId::List, a)) = arg_tys.first() {
-            if a.len() == 1 && !is_heap_ty(&a[0]) {
-                return Some("list.pop".to_string());
-            }
-        }
-        return Some("list.pop_x".to_string());
-    }
-    // `list.zip` keys on BOTH sources: scalar/scalar → flat pairs; FLAT-heap/FLAT-heap
-    // (String or List[scalar] each side — matrix rows) → the rc-share variant (the
-    // call-site `DropListStrStr` releases both acquired refs); anything else walls.
-    if func == "zip" && arg_tys.len() == 2 {
-        let elem = |t: &Ty| match t {
-            Ty::Applied(TypeConstructorId::List, a) if a.len() == 1 => Some(a[0].clone()),
-            _ => None,
-        };
-        if let (Some(ea), Some(eb)) = (elem(&arg_tys[0]), elem(&arg_tys[1])) {
-            let flat_heap = |t: &Ty| matches!(t, Ty::String)
-                || matches!(t, Ty::Applied(TypeConstructorId::List, b)
-                    if b.len() == 1 && !is_heap_ty(&b[0]));
-            if !is_heap_ty(&ea) && !is_heap_ty(&eb) {
-                return Some("list.zip".to_string());
-            }
-            if flat_heap(&ea) && flat_heap(&eb) {
-                return Some("list.zip_rc".to_string());
-            }
-            // MIXED scalar/flat-heap: co-own only the heap side (`_sh`/`_hs`).
-            if !is_heap_ty(&ea) && flat_heap(&eb) {
-                return Some("list.zip_sh".to_string());
-            }
-            if flat_heap(&ea) && !is_heap_ty(&eb) {
-                return Some("list.zip_hs".to_string());
-            }
-            return Some("list.zip_h".to_string());
-        }
-    }
-    // `list.repeat` over a HEAP element (`list.repeat(h, n_rep)` where `h: Matrix` — the
-    // nn repeat_kv GQA duplication): each result slot must CO-OWN the element (rc_inc per
-    // copy) — the scalar impl's raw alias would make every slot an uncounted owner the
-    // recursive result drop double-frees.
-    if func == "repeat" {
-        if let Some(t) = arg_tys.first() {
-            if is_heap_ty(t) {
-                return Some("list.repeat_rc".to_string());
-            }
-        }
-    }
-    // `list.flatten` over HEAP-element sublists: the copied slots are handles the
-    // result must CO-OWN — route to the rc_inc-on-copy variant (the scalar variant's
-    // raw copy would make the result a second uncounted owner = a double free).
-    if func == "flatten" {
-        if let Ty::Applied(TypeConstructorId::List, inner) = result_ty {
-            if inner.len() == 1 && is_heap_ty(&inner[0]) {
-                return Some("list.flatten_rc".to_string());
-            }
-        }
-    }
-    None
-}
-
-fn list_call_name_ordering(func: &str, arg_tys: &[Ty], result_ty: &Ty) -> Option<String> {
-    let _ = result_ty; // unused (kept for the shared `.or_else()` router signature)
-    if matches!(func, "sort" | "min" | "max") {
-        if let Some(r) = list_call_name_sort_min_max(func, arg_tys) {
-            return Some(r);
-        }
-    }
-    if func == "sort_by" {
-        if let Some(r) = list_call_name_sort_by(arg_tys) {
-            return Some(r);
-        }
-    }
-    None
-}
-
-/// The `sort`/`min`/`max` arm of [`list_call_name_ordering`] — verbatim move.
-fn list_call_name_sort_min_max(func: &str, arg_tys: &[Ty]) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // List[Float] ordering uses IEEE-754 totalOrder (f64::total_cmp), NOT a signed-int slot
-    // compare. Float is SCALAR (is_heap_ty false), so the heap routes below never fire for it —
-    // route sort/min/max explicitly on the element being Ty::Float (C-055). sort_by keys on the
-    // CLOSURE (arg 1) RETURN type being Float — the element list may be any type (e.g. List[R]).
-    if let Some(Ty::Applied(TypeConstructorId::List, a)) = arg_tys.first() {
-        if a.len() == 1 && a[0] == Ty::Float {
-            return Some(format!("list.{func}_float"));
-        }
-        // list.sort/min/max over a List[String] compare by CONTENT, not the i64 handle the
-        // generic impls compare (→ arbitrary/handle order or wrong element, a silent bug).
-        // _str variants do a lexicographic byte compare.
-        if a.len() == 1 && matches!(a[0], Ty::String) {
-            return Some(format!("list.{func}_str"));
-        }
-        // COMPOUND elements (the C-053 lattice) — type-directed comparators
-        // (list_ord_compound.almd): scalar/String 2-tuples, List[Int/Bool],
-        // List[String], Option[scalar]. Float components are EXCLUDED (IEEE
-        // totalOrder ≠ an i64 slot compare). Any OTHER heap element walls (`_x`):
-        // the generic twin would compare handle addresses AND raw-copy without
-        // rc_inc — a silent wrong order plus a double-free, never acceptable.
-        if a.len() == 1 && is_heap_ty(&a[0]) {
-            let scalar_nf = |t: &Ty| matches!(t, Ty::Int | Ty::Bool);
-            let suffix = match &a[0] {
-                Ty::Tuple(ts) if ts.len() == 2 && scalar_nf(&ts[0]) && scalar_nf(&ts[1]) => {
-                    Some("tss")
-                }
-                Ty::Tuple(ts)
-                    if ts.len() == 2 && scalar_nf(&ts[0]) && matches!(ts[1], Ty::String) =>
-                {
-                    Some("tsstr")
-                }
-                Ty::Applied(TypeConstructorId::List, e)
-                    if e.len() == 1 && scalar_nf(&e[0]) =>
-                {
-                    Some("lint")
-                }
-                Ty::Applied(TypeConstructorId::List, e)
-                    if e.len() == 1 && matches!(e[0], Ty::String) =>
-                {
-                    Some("lstr")
-                }
-                Ty::Applied(TypeConstructorId::Option, o)
-                    if o.len() == 1 && scalar_nf(&o[0]) =>
-                {
-                    Some("oint")
-                }
-                _ => None,
-            };
-            return Some(match suffix {
-                // sort_lstr has no twin yet — only min/max are implemented for
-                // the List[String] element family; sort keeps the honest wall.
-                Some("lstr") if func == "sort" => "list.sort_x".to_string(),
-                Some(s) => format!("list.{func}_{s}"),
-                None => format!("list.{func}_x"),
-            });
-        }
-    }
-    None
-}
-
-/// The `sort_by` arm of [`list_call_name_ordering`] — verbatim move.
-fn list_call_name_sort_by(arg_tys: &[Ty]) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId;
-    if let Some(Ty::Fn { ret, .. }) = arg_tys.get(1) {
-        // A HEAP element (List[String]/List[R]) must be CO-OWNED by the
-        // result list (rc_inc per copied handle) — the raw-copy variants
-        // share without acquiring and the two recursive drops double-free.
-        let heap_elem = matches!(arg_tys.first(),
-            Some(Ty::Applied(TypeConstructorId::List, a)) if a.len() == 1 && is_heap_ty(&a[0]));
-        if **ret == Ty::Float {
-            return Some(if heap_elem {
-                "list.sort_by_float_rc".to_string()
-            } else {
-                "list.sort_by_float".to_string()
-            });
-        }
-        // A STRING-key sort_by over SCALAR elements routes to the cached-key
-        // stable twin (byte-lexicographic via string.cmp — #560/C-055). A HEAP
-        // element still walls (`_x`): the copied handles would need the rc_inc
-        // co-own leg the scalar twin doesn't carry.
-        if **ret == Ty::String {
-            if !heap_elem {
-                return Some("list.sort_by_str_key".to_string());
-            }
-            return Some("list.sort_by_str_key_x".to_string());
-        }
-        if heap_elem {
-            return Some("list.sort_by_rc".to_string());
-        }
-    }
-    None
-}
-
-fn list_call_name_transform(func: &str, arg_tys: &[Ty], result_ty: &Ty) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // `list.map` is the one combinator whose SOURCE and RESULT element reprs may DIFFER (the
-    // closure transforms the type). A heap RESULT over a SCALAR source (`float.to_string` over a
-    // List[Float], `int.to_string` over a List[Int]) must read the source slot as a raw i64
-    // scalar (load64), not as a String handle (load_str) — that is `map_s2h`; a heap result over
-    // a heap source is the all-String `map_str`.
-    if func == "map" {
-        if let Ty::Applied(TypeConstructorId::List, rargs) = result_ty {
-            if rargs.len() == 1 && is_heap_ty(&rargs[0]) {
-                let src_heap = matches!(
-                    arg_tys.first(),
-                    Some(Ty::Applied(TypeConstructorId::List, s)) if s.len() == 1 && is_heap_ty(&s[0])
-                );
-                return Some(if src_heap {
-                    "list.map_str".to_string()
-                } else {
-                    "list.map_s2h".to_string()
-                });
-            }
-        }
-    }
-    // `list.enumerate` over a List[String] → `list.enumerate_str` (result List[(Int, String)]).
-    // Keyed on the SOURCE arg being List[String] (the yaml `lines |> list.enumerate` shape).
-    if func == "enumerate" {
-        if let Some(Ty::Applied(TypeConstructorId::List, s)) = arg_tys.first() {
-            if s.len() == 1 && matches!(s[0], Ty::String) {
-                return Some("list.enumerate_str".to_string());
-            }
-        }
-    }
-    // chunk/windows/window (build a NESTED List[List[heap]] whose recursive drop is a separate
-    // gap) and the HIGHER-ORDER take_while/drop_while/reduce over a HEAP element (String/Value):
-    // the generic i64 self-host impls copy element handles WITHOUT rc_inc → a DOUBLE-FREE at
-    // scope end (the result and source both free the shared handle), and the HO closure ABI is
-    // i64-scalar so a String/Value i32 handle mismatches the indirect call. Both are memory-
-    // safety bugs the prim-region ownership cert cannot see (it treats prim rc as a no-op), so
-    // they slipped past corpus-wall and trapped only at runtime. Route to an UNREGISTERED name →
-    // render walls cleanly (a controlled reject, never a miscompile or double-free) until the
-    // rc-correct heap variants land. Scalar element lists (Int/Float/Bool) are unaffected.
-    // take_while/drop_while over a List[String] now have rc-correct _str variants (each kept
-    // String DEEP-COPIED so result + source can both drop without a double-free); route to them
-    // BEFORE the heap-element wall below. (List[Value] / the other combinators still wall.)
-    // `reduce` over a List[String] joins them too: `list.reduce_str` is the self-host tail-`if`
-    // (None on empty, else Some) whose loop is the proven closure-call heap-accumulator (#738 Path
-    // C). Routed to the rc-correct _str variant BEFORE the heap-element wall.
-    if matches!(func, "take_while" | "drop_while" | "chunk" | "windows" | "window" | "reduce") {
-        if let Some(Ty::Applied(TypeConstructorId::List, s)) = arg_tys.first() {
-            if s.len() == 1 && matches!(s[0], Ty::String) {
-                return Some(format!("list.{func}_str"));
-            }
-        }
-    }
-    if matches!(func, "chunk" | "windows" | "window" | "take_while" | "drop_while" | "reduce") {
-        if let Some(Ty::Applied(TypeConstructorId::List, s)) = arg_tys.first() {
-            if s.len() == 1 && is_heap_ty(&s[0]) {
-                return Some(format!("list.{func}_heapelem"));
-            }
-        }
-    }
-    None
-}
-
-fn list_call_name_modifiers(func: &str, arg_tys: &[Ty], result_ty: &Ty) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // `list.set` over a List[String] → `list.set_str` (the val is a String HANDLE, not an i64
-    // Int — the generic list.set's i64 val param mismatches; set_str rc-copies + co-owns). The
-    // yaml `list.set(lines, dp, …)` shape.
-    // The heap-element list MODIFIERS (set/insert/remove_at/swap/update) over a List[String]/
-    // List[Value]: the generic i64 impls copy slots WITHOUT rc_inc (→ double-free when
-    // result+source both drop) and i32/i64-mismatch on a typed element (set/insert's `x`; for
-    // update the CLOSURE — a (String/Value)->same lambda renders an i32 handle RESULT while the
-    // generic f: (Int)->Int call_indirect expects i64 → runtime type-mismatch trap, #736). The
-    // _str/_value variants rc-copy each element to co-own and recursively free any replaced
-    // element. The element type is the first arg's List parameter.
-    if matches!(func, "set" | "insert" | "remove_at" | "swap" | "update") {
-        if let Some(Ty::Applied(TypeConstructorId::List, s)) = arg_tys.first() {
-            if s.len() == 1 && matches!(s[0], Ty::String) {
-                return Some(format!("list.{func}_str"));
-            }
-            if s.len() == 1 && is_value_ty(&s[0]) {
-                return Some(format!("list.{func}_value"));
-            }
-        }
-    }
-    // The element-PRESERVING List[heap]-returning combinators (source elem == result elem).
-    // slice joined for the same reason take/drop did: the Int twin's raw i64 element copy
-    // aliased String handles un-owned — a scope-end double free (fuzz seed-20260718 index 680).
-    if matches!(func, "filter" | "reverse" | "take" | "drop" | "slice" | "unique" | "dedup" | "intersperse") {
-        if let Ty::Applied(TypeConstructorId::List, args) = result_ty {
-            // A List[List[String]] result element is itself a heap list — the `_str` deep-copy
-            // (string.repeat) would read its length word as a byte count. take/drop SHARE the inner
-            // lists by handle via the `_liststr` variant; the other combinators are a later brick.
-            if args.len() == 1
-                && matches!(func, "take" | "drop")
-                && matches!(&args[0], Ty::Applied(TypeConstructorId::List, e)
-                    if e.len() == 1 && matches!(e[0], Ty::String))
-            {
-                return Some(format!("list.{func}_liststr"));
-            }
-            if args.len() == 1 && matches!(args[0], Ty::String) {
-                return Some(format!("list.{func}_str"));
-            }
-            // A NON-String heap element (a custom variant / record / Value handle):
-            // the `_str` deep-copy would read the block's length word as a byte
-            // count (garbage handles — the closures_and_variants UAF). `filter`
-            // routes to the rc-sharing `_rc` variant; the other combinators have
-            // no handle-sharing self-host yet, so route them to an UNREGISTERED
-            // `_hshare` name — the render walls it cleanly (the fold_hacc
-            // precedent), never a miscompile.
-            if args.len() == 1 && is_heap_ty(&args[0]) {
-                if func == "filter" {
-                    return Some("list.filter_rc".to_string());
-                }
-                // unique/dedup COMPARE elements — the `_hshare` slot-wise eq is exact
-                // ONLY for a flat scalar block (all-scalar tuple / List[scalar]). A
-                // String-BEARING element (record/tuple with a String slot) would
-                // compare the String HANDLE — two fresh identical strings mis-compare
-                // as distinct (CONFIRMED: unique over an inferred record list returned
-                // 3, native 2). The String-field record case routes to the generated
-                // `__krec_*` twin BEFORE this router; everything else walls.
-                if matches!(func, "unique" | "dedup")
-                    && !is_flat_scalar_block_ty(&args[0])
-                {
-                    return Some(format!("list.{func}_x"));
-                }
-                return Some(format!("list.{func}_hshare"));
-            }
-        }
-    }
-    None
-}
-
-fn list_call_name_accessors(func: &str, arg_tys: &[Ty], result_ty: &Ty) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // Element-RETURNING accessors / search over a List[heap] (the result is an Option[heap]):
-    // get/first/last (positional) + find (predicate higher-order).
-    if matches!(func, "get" | "first" | "last" | "find") {
-        if let Ty::Applied(TypeConstructorId::Option, args) = result_ty {
-            // A List[Value] element is a dynamic Value, NOT a String — the `_str` variant DEEP-
-            // COPIES via `string.repeat` (corrupting an Object to {}). Route get/first/last to the
-            // Value accessor, which SHARES the element (rc_inc, like value.get's Ok). (find's
-            // closure-keyed Value form is a later brick — only the positional accessors here.)
-            if args.len() == 1 && is_value_ty(&args[0]) && matches!(func, "get" | "first" | "last")
-            {
-                return Some(format!("list.{func}_value"));
-            }
-            // `find` over a `List[(Int, String)]` (the `enumerate |> find(closure)` shape): the
-            // element is a tuple, NOT a String — route to `find_int_str`, which loads each element
-            // as a TUPLE HANDLE and hands it to the predicate closure (reading e.g. `e.1`). The
-            // `_str` fallback would load it as a String handle (garbage).
-            if func == "find"
-                && args.len() == 1
-                && matches!(&args[0],
-                    Ty::Tuple(tys) if tys.len() == 2 && matches!(tys[0], Ty::Int) && matches!(tys[1], Ty::String))
-            {
-                return Some("list.find_int_str".to_string());
-            }
-            // A `List[List[String]]` element is itself a heap list, NOT a String — the `_str`
-            // variant would DEEP-COPY it via `string.repeat`, reading the inner list's length word
-            // as a byte count (garbage). Route to the handle-SHARE `_liststr` accessor (the
-            // `List[String]` analogue of `_value`); the inner list is co-owned, dropped DropListStr.
-            if args.len() == 1
-                && matches!(func, "get" | "first" | "last")
-                && matches!(&args[0], Ty::Applied(TypeConstructorId::List, e)
-                    if e.len() == 1 && matches!(e[0], Ty::String))
-            {
-                return Some(format!("list.{func}_liststr"));
-            }
-            // A RECORD/aggregate element (`list.get(vars, idx)` over `List[EnvVar]`
-            // — porta dedup_env): SHARE the element handle exactly like the
-            // Value/liststr accessors (the record stays owned by the list; the
-            // `_str` deep copy would read its block as string bytes). The `_value`
-            // impl is layout-identical (load_handle + Some), so reuse it. Keyed by
-            // elimination (this is a free fn, no layout registry): a heap element
-            // that is NOT a String/List/Value is a nominal record/tuple/variant
-            // block — all of which are single-handle share-safe.
-            if args.len() == 1
-                && matches!(func, "get" | "first" | "last")
-                && is_heap_ty(&args[0])
-                && !matches!(args[0], Ty::String)
-                && !is_value_ty(&args[0])
-                && !matches!(&args[0], Ty::Applied(TypeConstructorId::List | TypeConstructorId::Map | TypeConstructorId::Set, _))
-            {
-                return Some(format!("list.{func}_value"));
-            }
-            // `find`'s Some() result is a DEEP COPY of the found element (`string.repeat`,
-            // list_find_str's `__lfs_some`) — correct only for an actual String element; any
-            // other heap type here (a flat scalar tuple, a record) falls through to a WALL
-            // rather than corrupting the copy (the same class of bug fixed above for
-            // contains/index_of — `find` just hasn't grown a `_hshare` copy variant yet).
-            if args.len() == 1 && matches!(args[0], Ty::String) {
-                return Some(format!("list.{func}_str"));
-            }
-            // Any remaining heap-element shape (find over Value/List/Map/Set; get/first/last
-            // over a non-String heap List element) has no covering arm above — route to a
-            // deliberately UNREGISTERED `_x` name. The bare `list.{func}` name is NOT a safe
-            // fallback here: it links against the Int-typed generic self-host (list_search.almd
-            // et al), which silently "succeeds" via raw i64-slot / pointer-identity comparison
-            // instead of refusing to link (the fallthrough danger confirmed by probe elsewhere
-            // in this dispatch — see the contains/index_of comment above).
-            // get/first/last over a FLAT scalar-slot block element (List[scalar] /
-            // all-scalar tuple): the handle-SHARING variant (Some payload rc_inc'd
-            // by the Some-ctor Dup) — the deep-copy `_str` corrupts these shapes.
-            if args.len() == 1
-                && matches!(func, "get" | "first" | "last")
-                && is_flat_scalar_block_ty(&args[0])
-            {
-                return Some(format!("list.{func}_hshare"));
-            }
-            if args.len() == 1 && is_heap_ty(&args[0]) {
-                return Some(format!("list.{func}_x"));
-            }
-        }
-    }
-    // `list.partition` keys on its element: scalar → the raw-copy self-host; a FLAT
-    // heap element (String / List[scalar] / scalar tuple) → the co-owning `_rc`
-    // variant; any richer element walls (`_x` — no recursive-co-own variant yet).
-    if func == "partition" {
-        if let Some(Ty::Applied(TypeConstructorId::List, a)) = arg_tys.first() {
-            if a.len() == 1 {
-                if !is_heap_ty(&a[0]) {
-                    return Some("list.partition".to_string());
-                }
-                if matches!(a[0], Ty::String) || is_flat_scalar_block_ty(&a[0]) {
-                    return Some("list.partition_rc".to_string());
-                }
-                return Some("list.partition_x".to_string());
-            }
-        }
-    }
-    // get_or returns the ELEMENT directly (not an Option). Over a List[heap] it must return
-    // an i32 handle (a deep copy), so it is keyed on the heap RESULT being the element type.
-    if func == "get_or" && is_heap_ty(result_ty) {
-        // A Value element SHARES (the _str deep copy corrupts a Value block —
-        // json_gltf_walk's `list.get_or(meshes, 0, json.null())` returned "?").
-        if is_value_ty(result_ty) {
-            return Some("list.get_or_value".to_string());
-        }
-        if matches!(result_ty, Ty::String) {
-            return Some("list.get_or_str".to_string());
-        }
-        // No handle-sharing self-host for other heap elements yet — an
-        // UNREGISTERED name walls cleanly (the fold_hacc precedent).
-        return Some("list.get_or_hshare".to_string());
-    }
-    // SUBJECT-keyed (arg 0) over a List[heap], where the result is scalar (Bool/Int/Option[Int])
-    // so it can't be keyed on the result type: search (contains/index_of) does an ELEMENT-
-    // EQUALITY comparison — String routes to the byte-eq `_str` family (correct only for actual
-    // String elements: __str_eq reads the length FIELD as a byte count); a flat scalar-slot
-    // block (all-scalar tuple / List[scalar]) routes to the slot-wise `_hshare` family (B32's
-    // `__uh_eq`, which reads length as an ELEMENT count and compares raw i64 slots — exact for
-    // this shape). Any OTHER heap element (record, Value, nested heap list, String-bearing
-    // tuple) has no correct comparison variant yet — falls through to a WALL (never routed to
-    // `_str`, which would silently produce WRONG results: a tuple/list `len` misread as a byte
-    // count truncates the compare to its first ~2 bytes, a confirmed false-positive collision).
-    // NOTE: falling through to the bare `list.contains`/`list.index_of` name here is NOT a
-    // safe wall — it links against the Int-typed generic (list_search.almd), which silently
-    // "succeeds" (raw i64-slot compare, i.e. POINTER-IDENTITY on a heap handle — the exact
-    // OLD C-015 bug) rather than refusing to link (CONFIRMED by probe: a record-element
-    // `list.contains` produced invalid wasm via this path, and a tuple-element `set.from_list`
-    // sharing the same generic-fallthrough shape silently mis-deduped). Any excluded heap
-    // element must route to a deliberately UNREGISTERED name (`_x`, the established
-    // wall-suffix convention) so the render step's unlinked-call check catches it.
-    if matches!(func, "contains" | "index_of") {
-        if let Some(Ty::Applied(TypeConstructorId::List, a)) = arg_tys.first() {
-            if a.len() == 1 && is_heap_ty(&a[0]) {
-                if matches!(a[0], Ty::String) {
-                    return Some(format!("list.{func}_str"));
-                }
-                if is_flat_scalar_block_ty(&a[0]) {
-                    return Some(format!("list.{func}_hshare"));
-                }
-                return Some(format!("list.{func}_x"));
-            }
-        }
-    }
-    // all/any/count/fold are pure closure-passthrough (each element handle is loaded and handed
-    // to the predicate/accumulator closure — no internal equality compare or deep copy), so the
-    // byte-eq `_str` family's ITERATION shape is safe for ANY heap element type, not just String.
-    if matches!(func, "all" | "any" | "count" | "fold") {
-        if let Some(Ty::Applied(TypeConstructorId::List, a)) = arg_tys.first() {
-            if a.len() == 1 && is_heap_ty(&a[0]) {
-                return Some(format!("list.{func}_str"));
-            }
-        }
-    }
-    None
-}
-
-
-/// Every `set.*` typed-variant route (see the block comments).
-fn set_call_name(func: &str, arg_tys: &[Ty], result_ty: &Ty) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // `Set[heap]`-RETURNING constructors key on the RESULT element type; `set.to_list` over a
-    // `Set[heap]` returns a `List[heap]`; the predicate `set.contains` keys on its SUBJECT
-    // (arg 0) element type (its result is Bool). Every one of these funcs relies on `__str_eq`
-    // (byte-level String-layout equality, for membership/dedup) AND/OR `string.repeat` (a
-    // String-specific deep copy) internally (set_str.almd) — BOTH are unsound for a non-String
-    // heap element: `__str_eq` misreads a block's slot-count `len` as a byte count (a confirmed
-    // false-positive collision past the first ~2 bytes), and `string.repeat` would corrupt a
-    // tuple/record/list block's bytes. Restrict the `_str` route to an ACTUAL String element;
-    // any other heap element (tuple, record, nested list, Value) falls through to a WALL — no
-    // correct Set variant exists yet for those (the flat-scalar `_hshare` family list.contains
-    // just grew does not cover Set's dedup-on-build/algebra ops).
-    let result_elem_is_string = matches!(
-        result_ty,
-        Ty::Applied(TypeConstructorId::Set | TypeConstructorId::List, a)
-            if a.len() == 1 && matches!(a[0], Ty::String)
-    );
-    // RESULT-keyed: constructors / Set-returning algebra over heap elements. The bare
-    // `set.{func}` fallback is NOT a safe wall for a non-String heap element here — it links
-    // against the Int-typed generic (set_core.almd), which silently "succeeds" via raw i64-slot
-    // / pointer-identity comparison (the OLD C-015 bug) instead of refusing to link (CONFIRMED
-    // by probe: `set.from_list` over a `List[(Int,Int)]` silently mis-deduped, len 3 instead of
-    // 2, then trapped later). Route explicitly to the UNREGISTERED `_x` wall name instead.
-    let result_elem_is_heap = matches!(
-        result_ty,
-        Ty::Applied(TypeConstructorId::Set | TypeConstructorId::List, a)
-            if a.len() == 1 && is_heap_ty(&a[0])
-    );
-    // FLAT all-scalar TUPLE elements (`Set[(Int, Int)]` — the C-015 compound-eq
-    // class): each element normalizes through `__srec_key` (map_vkey.almd — the
-    // tuple block is physically the all-scalar record) into a backing `_str` set,
-    // so dedup/lookup are CONTENT equality. Build/lookup fns only — an
-    // iteration-returning fn (to_list) would surface the normalized strings and
-    // keeps the wall below.
-    {
-        let tup_set = |t: &Ty| {
-            matches!(t, Ty::Applied(TypeConstructorId::Set, a) if a.len() == 1
-                && matches!(&a[0], Ty::Tuple(ts)
-                    if !ts.is_empty() && ts.iter().all(|c| matches!(c, Ty::Int | Ty::Bool))))
-        };
-        if matches!(func, "from_list" | "insert" | "contains")
-            && (arg_tys.first().is_some_and(tup_set) || tup_set(result_ty))
-        {
-            return Some(format!("set.{func}_srec"));
-        }
-    }
-    if matches!(
-        func,
-        "from_list" | "to_list" | "union" | "intersection" | "difference"
-            | "new" | "insert" | "remove" | "symmetric_difference" | "filter"
-    ) && result_elem_is_heap
-    {
-        return Some(if result_elem_is_string {
-            format!("set.{func}_str")
-        } else {
-            format!("set.{func}_x")
-        });
-    }
-    // TYPE-CHANGING `set.map` to a HEAP element result (C-039's `set.map(si, (x) =>
-    // "n" + …)`): the generic's i64-result closure table type traps on the i32-result
-    // closure. A scalar subject → String result composes through the insert_str twin;
-    // any other heap result walls (`_x`). (A scalar RESULT over any subject keeps the
-    // generic — its dedup and closure table are i64-correct there.)
-    if func == "map" && result_elem_is_heap {
-        let subj_scalar = matches!(arg_tys.first(),
-            Some(Ty::Applied(TypeConstructorId::Set, a)) if a.len() == 1 && !is_heap_ty(&a[0]));
-        return Some(if result_elem_is_string && subj_scalar {
-            "set.map_i2s".to_string()
-        } else {
-            "set.map_x".to_string()
-        });
-    }
-    // `all`/`any`/`fold` are pure closure-passthrough (no internal eq compare or deep copy) —
-    // safe for any heap element type, same as the list-module analogue above.
-    let arg0_elem_is_heap = matches!(
-        arg_tys.first(),
-        Some(Ty::Applied(TypeConstructorId::Set, a)) if a.len() == 1 && is_heap_ty(&a[0])
-    );
-    if matches!(func, "all" | "any" | "fold") && arg0_elem_is_heap {
-        return Some(format!("set.{func}_str"));
-    }
-    // ARG-keyed eq/membership: a Bool-returning fn over a `Set[heap]` subject (arg 0) — String
-    // element only (the `__str_eq`/`__set_has_str` unsoundness above); any other heap element
-    // routes to the `_x` wall (the same fallthrough danger as the RESULT-keyed family).
-    let arg0_elem_is_string = matches!(
-        arg_tys.first(),
-        Some(Ty::Applied(TypeConstructorId::Set, a)) if a.len() == 1 && matches!(a[0], Ty::String)
-    );
-    if matches!(func, "contains" | "is_subset" | "is_disjoint" | "eq") && arg0_elem_is_heap {
-        return Some(if arg0_elem_is_string {
-            format!("set.{func}_str")
-        } else {
-            format!("set.{func}_x")
-        });
-    }
-    None
-}
-
-/// Every `map.*` typed-variant route: the skv entries/from_list fast paths,
-/// then the (key, value) repr-family table (see the block comments).
-fn map_call_name(func: &str, arg_tys: &[Ty], result_ty: &Ty, map_key_nullary: bool, map_key_scalar_rec: bool) -> Option<String> {
-    use almide_lang::types::constructor::TypeConstructorId;
-    // `map.entries` / `map.from_list` over the skv repr (`Map[String, scalar]` — the
-    // tokenizer vocab): route to the skv self-hosts; the all-String repr keeps its
-    // existing `map.entries_str`. Other reprs wall (unregistered).
-    if func == "entries" {
-        if let Some(Ty::Applied(TypeConstructorId::Map, a)) = arg_tys.first() {
-            if a.len() == 2 && matches!(a[0], Ty::String) && !is_heap_ty(&a[1]) {
-                return Some("map.entries_skv".to_string());
-            }
-        }
-    }
-    if func == "from_list" {
-        if let Some(Ty::Applied(TypeConstructorId::List, a)) = arg_tys.first() {
-            if a.len() == 1
-                && matches!(&a[0], Ty::Tuple(ts) if ts.len() == 2
-                    && matches!(ts[0], Ty::String) && !is_heap_ty(&ts[1]))
-            {
-                return Some("map.from_list_skv".to_string());
-            }
-        }
-    }
-    // A map's REPR is set by its (key, value) heap-ness, read from whichever Map type the call
-    // exposes: arg 0 (the SUBJECT of set/get/fold/filter/…) takes priority, else the RESULT
-    // (map.new() has no args). The two repr families:
-    //   key heap, value heap  → `_str` (map_str: interleaved all-String entries)
-    //   key heap, value scalar → `_skv` (map_skv: String keys + i64 values, serves
-    //                            Map[String,Int] AND Map[String,Float] — the value is one i64)
-    //   key scalar             → the plain map_core (Map[Int,Int]); a scalar-key heap-value map
-    //                            has no variant yet, so it falls through (walled by repr).
-    // The element-returning forms (get → Option[V], keys/values → List[elem]) read the same
-    // key/value reprs off the subject map (arg 0).
-    let map_kv = |ty: &Ty| match ty {
-        Ty::Applied(TypeConstructorId::Map, a) if a.len() == 2 => {
-            Some((is_heap_ty(&a[0]), is_heap_ty(&a[1])))
-        }
-        _ => None,
-    };
-    let kv = arg_tys
-        .first()
-        .and_then(&map_kv)
-        .or_else(|| map_kv(result_ty));
-    if let Some((key_heap, val_heap)) = kv {
-        // Each repr family routes ONLY the funcs its self-hosted variant file actually defines
-        // (an unlisted func keeps the plain name — never a dangling `_str`/`_skv` reference).
-        // Whether the VALUE type is exactly String — the `_str` variant stores
-        // interleaved all-String entries, so a heap-but-not-String value
-        // (`Map[String, List[Int]]`) must NOT route there (its deep-copy would
-        // read the list block as string bytes). It routes to an UNREGISTERED
-        // `_hval` name instead — a clean render wall, never invalid wasm.
-        let val_is_string = matches!(
-            arg_tys.first().or(Some(result_ty)),
-            Some(Ty::Applied(TypeConstructorId::Map, a)) if a.len() == 2 && matches!(a[1], Ty::String)
-        );
-        // Both the `_str` (all-String interleaved entries) and `_skv` (String key + i64 value)
-        // families do KEY EQUALITY via `__str_eq`/`__skv_eq` — a byte-level compare that
-        // misreads a non-String heap block's slot-count `len` as a byte count (the same
-        // confirmed false-positive-collision class fixed above for list/set). A non-String heap
-        // KEY (a tuple, record, nested list) must NOT route to either family — CONFIRMED via
-        // probe to currently produce INVALID WASM (an i32/i64 ABI-width mismatch on `map.set`
-        // with a tuple key), not silently wrong bytes, but still not the honest wall this repr
-        // gate is meant to guarantee. Gate both families on an ACTUAL String key.
-        let key_is_string = matches!(
-            arg_tys.first().or(Some(result_ty)),
-            Some(Ty::Applied(TypeConstructorId::Map, a)) if a.len() == 2 && matches!(a[0], Ty::String)
-        );
-        // The FLAT-heap-value class map_hval actually implements (List[scalar]).
-        let val_is_flat_list = matches!(
-            arg_tys.first().or(Some(result_ty)),
-            Some(Ty::Applied(TypeConstructorId::Map, a))
-                if a.len() == 2 && matches!(&a[1],
-                    Ty::Applied(TypeConstructorId::List, e) if e.len() == 1 && !is_heap_ty(&e[0]))
-        );
-        // The CLOSURE-value class (`Map[String, () -> Unit]` — the mclo family): the
-        // hval twins are handle-level (set stores + rc-shares the value handle, get/
-        // get_or share it back), so a closure block rides them unchanged; only the
-        // DROP routes differently (`$__drop_map_mclo` — the bind/arg registration
-        // sites key on `is_map_fn_ty`). `eq` is EXCLUDED (hval eq compares values
-        // STRUCTURALLY as List — closure identity has no such compare; it stays an
-        // honest wall).
-        let val_is_fn = matches!(
-            arg_tys.first().or(Some(result_ty)),
-            Some(Ty::Applied(TypeConstructorId::Map, a))
-                if a.len() == 2 && matches!(a[1], Ty::Fn { .. })
-        );
-        let variant = match (key_heap, val_heap) {
-            // `Map[String, List[scalar]]` — the implemented subset of the heap-value
-            // family (new/set/eq/len/contains/get/get_or; other funcs keep the
-            // unregistered wall name). `get`/`get_or` SHARE the stored value (the
-            // hshare discipline).
-            (true, true)
-                if val_is_flat_list
-                    && matches!(
-                        func,
-                        "new" | "set" | "eq" | "len" | "contains" | "get" | "get_or" | "keys"
-                    ) =>
-            {
-                Some("_hval")
-            }
-            (true, true)
-                if val_is_fn
-                    && matches!(func, "new" | "set" | "len" | "contains" | "get" | "get_or") =>
-            {
-                Some("_hval")
-            }
-            // `Map[String, <Fn>]` from_list (keyed on the RESULT — the arg is the pairs
-            // List): the hval pair-walk is handle-level (each value rc-shared in via
-            // `map_set_hval`), so a closure block rides it unchanged; the RESULT's
-            // type-driven drop routing (`is_map_fn_ty` → map_mclo) frees the values via
-            // `__drop_closure`.
-            (true, true)
-                if func == "from_list"
-                    && matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                        if a.len() == 2 && matches!(a[0], Ty::String)
-                            && matches!(a[1], Ty::Fn { .. })) =>
-            {
-                Some("_hval")
-            }
-            // `Map[String, List[Int]]` from_list / display (the map-of-lists literal):
-            // keyed on the RESULT/first-arg map; to_string_hval passes through
-            // verbatim (the B22 suffix guard).
-            // `Map[String, String]` from_list (the String-valued map literal): keyed on
-            // the RESULT type (from_list's first arg is the pairs List, not a Map).
-            (true, true)
-                if func == "from_list"
-                    && matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                        if a.len() == 2
-                            && matches!(a[0], Ty::String)
-                            && matches!(a[1], Ty::String)) =>
-            {
-                Some("_str")
-            }
-            (true, true)
-                if !val_is_string
-                    && func == "from_list"
-                    && matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                        if a.len() == 2
-                            && matches!(a[0], Ty::String)
-                            && matches!(&a[1], Ty::Applied(TypeConstructorId::List, b)
-                                if b.len() == 1 && matches!(b[0], Ty::Int))) =>
-            {
-                Some("_hval")
-            }
-            // `Map[String, Map[String, String]]` get_or / from_list — the msv family
-            // (map_fold_heap_acc's nested-map literal + get_or default).
-            (true, true)
-                if func == "get_or"
-                    && matches!(arg_tys.first(), Some(Ty::Applied(TypeConstructorId::Map, a))
-                        if a.len() == 2 && matches!(a[0], Ty::String)
-                            && matches!(&a[1], Ty::Applied(TypeConstructorId::Map, b)
-                                if b.len() == 2
-                                    && matches!(b[0], Ty::String)
-                                    && matches!(b[1], Ty::String))) =>
-            {
-                Some("_msv")
-            }
-            (true, true)
-                if func == "from_list"
-                    && matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                        if a.len() == 2 && matches!(a[0], Ty::String)
-                            && matches!(&a[1], Ty::Applied(TypeConstructorId::Map, b)
-                                if b.len() == 2
-                                    && matches!(b[0], Ty::String)
-                                    && matches!(b[1], Ty::String))) =>
-            {
-                Some("_msv")
-            }
-            // `Map[String, List[Option[Int]]]` from_list — the mlo family
-            // (compound_repr_interp's `deep` inner-map literal).
-            (true, true)
-                if func == "from_list"
-                    && crate::lower::is_map_mlo_ty(result_ty) =>
-            {
-                Some("_mlo")
-            }
-            (true, true) if func == "to_string_hval" => Some(""),
-            // An ALREADY-SUFFIXED mlo display (`map.to_string_mlo` from the interp
-            // leaf) — pass through verbatim (re-suffixing would fabricate
-            // `to_string_mlo_hval_wall`).
-            (true, true) if func == "to_string_mlo" => Some(""),
-            // `map.from_list` over a NAMED-value map (`["o": Point{..}]` / `["a":
-            // Circle(3.0)]` — the desugared map literal): construction is handle-level
-            // (the `_str` family's pair copy + co-own rc_inc works for ANY heap value
-            // slot); the RESULT's type-driven drop routing (`map_named_value_drop`)
-            // decides the correct sweep, and an unadmitted value type walls THERE —
-            // never a leaky flat link here.
-            (true, true)
-                if func == "from_list"
-                    && matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                        if a.len() == 2 && matches!(a[0], Ty::String)
-                            && matches!(a[1], Ty::Named(..))) =>
-            {
-                Some("_hobj")
-            }
-            // An ALL-SCALAR record key with a String value (`Map[Color, String]` —
-            // the hash_protocol deriving-Hash shape): the key normalizes to the
-            // comma-joined decimal string of its slots (map_vkey.almd's _srec
-            // family — content identity ⇔ string identity for Int/Bool fields).
-            (true, true)
-                if map_key_scalar_rec
-                    && matches!(func, "from_list" | "get")
-                    && {
-                        let str_val = |t: &Ty| {
-                            matches!(t, Ty::Applied(TypeConstructorId::Map, a)
-                                if a.len() == 2 && matches!(a[1], Ty::String))
-                        };
-                        arg_tys.first().is_some_and(str_val) || str_val(result_ty)
-                    } =>
-            {
-                Some("_srec")
-            }
-            // `map.entries` over the hval-TUPLE flavor (`Map[String, (Int, Int)]` —
-            // the C-039 tuple-valued map.map result): the typechange twin.
-            (true, true)
-                if func == "entries"
-                    && matches!(arg_tys.first(), Some(Ty::Applied(TypeConstructorId::Map, a))
-                        if a.len() == 2 && matches!(a[0], Ty::String)
-                            && matches!(&a[1], Ty::Tuple(ts)
-                                if !ts.is_empty() && ts.iter().all(|c| !is_heap_ty(c)))) =>
-            {
-                Some("_hvalt")
-            }
-            // TYPE-CHANGING `map.map` str → skv (`map.map(ms, (v) => string.len(v))`
-            // — C-039): the result value narrows to a scalar.
-            (true, true)
-                if key_is_string
-                    && func == "map"
-                    && matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                        if a.len() == 2 && matches!(a[0], Ty::String) && !is_heap_ty(&a[1])) =>
-            {
-                Some("_str2skv")
-            }
-            (true, true) if !val_is_string => Some("_hval_wall"),
-            // A fn OUTSIDE the family's implemented list must fall to an UNREGISTERED
-            // wall suffix, NEVER to `None`: a `None` here returns the BARE `map.{func}`
-            // name, which links the scalar-key map_core generic against the string-key
-            // 16-byte-stride layout — raw i64 slot copies of STRING handles with no
-            // rc_inc, so the result map aliases its inputs' keys unowned and scope-end
-            // double-frees them (`map.merge` on `Map[String, Int]` trapped the rc_dec
-            // sentinel on the verified default — the #790 map-merge row).
-            // `to_string` (the `${map}` interp display) links the BARE self-host
-            // `map.to_string` — the pre-wall behavior this family's wall must not
-            // mangle (map_interp_self_hosts_via_keys_values pins it).
-            (true, true) | (true, false) if key_is_string && func == "to_string" => Some(""),
-            (true, true) if key_is_string => Some(
-                if matches!(
-                    func,
-                    "new" | "set" | "remove" | "merge" | "update" | "filter" | "get" | "keys"
-                        | "values" | "len" | "is_empty" | "contains" | "all" | "any" | "count"
-                        | "fold" | "entries"
-                ) {
-                    "_str"
-                } else {
-                    "_str_wall"
-                },
-            ),
-            // TYPE-CHANGING `map.map` skv → str (`map.map(mi, (v) => int.to_string(v)
-            // + "!")` — C-039): String values in the result → the typechange twin.
-            (true, false)
-                if key_is_string
-                    && func == "map"
-                    && matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                        if a.len() == 2 && matches!(a[0], Ty::String)
-                            && matches!(a[1], Ty::String)) =>
-            {
-                Some("_skv2str")
-            }
-            // TYPE-CHANGING `map.map` skv → hval-TUPLE (`map.map(mi, (v) => (v, v*v))`
-            // — C-039): all-scalar tuple values → the raw skv-split build twin.
-            (true, false)
-                if key_is_string
-                    && func == "map"
-                    && matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                        if a.len() == 2 && matches!(a[0], Ty::String)
-                            && matches!(&a[1], Ty::Tuple(ts)
-                                if !ts.is_empty() && ts.iter().all(|c| !is_heap_ty(c)))) =>
-            {
-                Some("_skv2hvalt")
-            }
-            // `map.map` transforms the VALUES — the skv impl is scalar-value in AND out, so
-            // it also needs the RESULT map's value scalar (a `(v) => int.to_string(v)` maps
-            // into the `_str` repr — no skv form; wall it rather than mislink).
-            (true, false)
-                if key_is_string
-                    && func == "map"
-                    && !matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                        if a.len() == 2 && !is_heap_ty(&a[1])) =>
-            {
-                Some("_skv_wall")
-            }
-            // An ALREADY-SUFFIXED skv-repr display (`map.to_string_sb` from the interp
-            // leaf — `${Map[String, Bool]}`) — pass through verbatim (re-suffixing
-            // would fabricate `to_string_sb_skv_wall`, the to_string_hval lesson).
-            (true, false) if func == "to_string_sb" => Some(""),
-            (true, false) if key_is_string => Some(
-                if matches!(
-                    func,
-                    "new" | "set" | "remove" | "filter" | "get" | "get_or" | "keys" | "values"
-                        | "len" | "is_empty" | "contains" | "all" | "any" | "count" | "fold"
-                        | "eq" | "find" | "update" | "merge" | "map"
-                ) {
-                    "_skv"
-                } else {
-                    "_skv_wall"
-                },
-            ),
-            // A NULLARY-ONLY variant key with a SCALAR value (`Map[Direction, Int]` —
-            // the hash_protocol deriving-Hash shape): the key's identity IS its tag,
-            // so normalize to a raw i64 key (map_vkey.almd's from_list/get). Other
-            // funcs and heap values keep the honest wall below.
-            (true, false)
-                if map_key_nullary
-                    && matches!(func, "from_list" | "get")
-                    && {
-                        // from_list's FIRST arg is the pairs List — probe the
-                        // RESULT type too (the is_ivh OR discipline).
-                        let scalar_val = |t: &Ty| {
-                            matches!(t, Ty::Applied(TypeConstructorId::Map, a)
-                                if a.len() == 2 && !is_heap_ty(&a[1]))
-                        };
-                        arg_tys.first().is_some_and(scalar_val) || scalar_val(result_ty)
-                    } =>
-            {
-                Some("_vtag")
-            }
-            // FLAT all-scalar TUPLE key with a String value (`Map[(Int, Int), String]`
-            // — C-015): the key normalizes through `__srec_key` (the tuple block IS the
-            // all-scalar record physically) into the backing `_str` map — content
-            // identity ⇔ string identity. `len` reads the backing str-flavor header
-            // directly; the lookup/build set routes `_srec`.
-            (true, true) | (true, false)
-                if {
-                    let tup_key_str_val = |t: &Ty| {
-                        matches!(t, Ty::Applied(TypeConstructorId::Map, a) if a.len() == 2
-                            && matches!(&a[0], Ty::Tuple(ts)
-                                if !ts.is_empty()
-                                    && ts.iter().all(|c| matches!(c, Ty::Int | Ty::Bool)))
-                            && matches!(a[1], Ty::String))
-                    };
-                    (arg_tys.first().is_some_and(tup_key_str_val)
-                        || tup_key_str_val(result_ty))
-                        && matches!(func, "from_list" | "get" | "set" | "contains" | "len")
-                } =>
-            {
-                if func == "len" {
-                    return Some("map.len_str".to_string());
-                }
-                Some("_srec")
-            }
-            // A non-String heap KEY (tuple/record/nested list) reaching here has no correct
-            // variant — route to an explicit UNREGISTERED wall name rather than falling through
-            // to the bare `map.{func}` name, which links against the scalar-key map_core generic
-            // and produces INVALID WASM (an i32/i64 ABI-width mismatch, CONFIRMED by probe) —
-            // a crash, not the honest compile-time wall this repr gate exists to guarantee.
-            (true, true) | (true, false) => Some("_key_wall"),
-            // `Map[Int, String]` — the implemented scalar-key/heap-value variant
-            // (new/set/eq). Other funcs, and other heap value types, keep an
-            // UNREGISTERED wall name (never the plain Map[Int,Int] i64-slot link
-            // that emitted invalid wasm — map_set_eq's original failure).
-            // An ALREADY-SUFFIXED synthesized display call (`map.to_string_ivh` from
-            // the interp leaf table) — pass through verbatim (re-suffixing would
-            // fabricate `to_string_ivh_ivh_wall`).
-            (false, true) if func == "to_string_ivh" => Some(""),
-            (false, true)
-                if {
-                    // `from_list`'s FIRST arg is the pairs List, not the Map — key
-                    // its admission on the RESULT type instead (either probe works
-                    // for the Map-first fns).
-                    // A Bool key is the SAME raw i64 slot (0/1) — the ivh find's
-                    // i64 compare serves it verbatim, so admit it alongside Int.
-                    let is_ivh = |t: &Ty| {
-                        matches!(t, Ty::Applied(TypeConstructorId::Map, a)
-                            if a.len() == 2
-                                && matches!(a[0], Ty::Int | Ty::Bool)
-                                && matches!(a[1], Ty::String))
-                    };
-                    (arg_tys.first().is_some_and(is_ivh) || is_ivh(result_ty))
-                        && matches!(func, "new" | "set" | "eq" | "from_list" | "len" | "get")
-                } =>
-            {
-                Some("_ivh")
-            }
-            // TYPE-CHANGING `map.map` ivh → core (`map.map(mik, (v) => string.len(v))`
-            // — C-039): the String values narrow to scalars.
-            (false, true)
-                if func == "map"
-                    && matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                        if a.len() == 2
-                            && matches!(a[0], Ty::Int | Ty::Bool)
-                            && !is_heap_ty(&a[1])) =>
-            {
-                Some("_ivh2core")
-            }
-            (false, true) => Some("_ivh_wall"),
-            // `Map[Int, Float]` from_list (the float_map literal): ONE scalar-KV impl
-            // (map_if.almd) — the paired i64 slots carry f64 bits verbatim. Display
-            // routes separately (the interp table's to_string_if). Other scalar-scalar
-            // maps keep the plain (unlinked) name — walls cleanly.
-            (false, false)
-                if func == "from_list"
-                    && matches!(result_ty, Ty::Applied(TypeConstructorId::Map, a)
-                        if a.len() == 2
-                            && matches!(a[0], Ty::Int)
-                            && matches!(a[1], Ty::Float)) =>
-            {
-                Some("_if")
-            }
-            (false, false) if func == "to_string_if" => Some(""),
-            _ => None,
-        };
-        if let Some(suffix) = variant {
-            return Some(format!("map.{func}{suffix}"));
-        }
-    }
-    None
-}
-
-
